### 2016-11-23 cefbab7
improved unit test and robustified getJobPars (fixes #23)
