+++ tests/testthat/test_addExperiments.R
@@ -21,33 +21,43 @@ test_that("addExperiments handles parameters correctly", {
 test_that("addExperiments creates default designs", {
   reg = makeExperimentRegistry(file.dir = NA, make.default = FALSE)
   prob = addProblem(reg = reg, "p1", data = iris)
+  prob = addProblem(reg = reg, "p2", data = cars)
   algo = addAlgorithm(reg = reg, "a1", fun = function(job, data, instance) nrow(data))
   algo = addAlgorithm(reg = reg, "a2", fun = function(job, data, instance) ncol(data))
   ids = addExperiments(reg = reg)
-  expect_equal(findExperiments(reg = reg)$job.id, 1:2)
-  expect_equal(as.character(reg$defs$problem), c("p1", "p1"))
-  expect_equal(as.character(reg$defs$algorithm), c("a1", "a2"))
+  expect_equal(findExperiments(reg = reg)$job.id, 1:4)
+  expect_equal(as.character(reg$defs$problem), rep(c("p1", "p2"), each = 2))
+  expect_equal(as.character(reg$defs$algorithm), rep(c("a1", "a2"), times = 2))
 })
 
 test_that("addExperiments / user provided designs", {
   reg = makeExperimentRegistry(file.dir = NA, make.default = FALSE)
-  prob = addProblem(reg = reg, "p1", data = iris, fun = function(...) list(...))
-  algo = addAlgorithm(reg = reg, "a1", fun = function(...) list(...))
-  algo = addAlgorithm(reg = reg, "a2", fun = function(...) ncol(data))
+  addProblem(reg = reg, "p1", data = iris, fun = function(...) list(...))
+  addProblem(reg = reg, "p2", data = cars, fun = function(...) list(...))
+  addAlgorithm(reg = reg, "a1", fun = function(...) list(...))
+  addAlgorithm(reg = reg, "a2", fun = function(...) ncol(data))
   prob.designs = list(p1 = data.table(a = 1, b = 2:4))
   algo.designs = list(a1 = data.table(c = 3:8), a2 = data.table())
   repls = 1
   ids = addExperiments(reg = reg, prob.designs = prob.designs, algo.designs = algo.designs, combine = "bind")
   expect_data_table(ids, nrow = 9, key = "job.id")
   pars = getJobPars(reg = reg)
-  expect_equal(as.character(unique(pars$problem)), "p1")
-  expect_set_equal(as.character(unique(pars$algorithm)), c("a1", "a2"))
+  expect_set_equal(as.character(pars$problem), "p1")
+  expect_set_equal(as.character(pars$algorithm), c("a1", "a2"))
   expect_equal(pars$a, rep(1L, 9))
   expect_equal(pars$b, rep(2:4, 3))
   expect_equal(pars$c, c(3:8, rep(NA, 3)))
-
   expect_error(addExperiments(reg = reg, prob.designs = list(p1 = data.table(job = 2))), "reserved keyword 'job'")
   expect_error(addExperiments(reg = reg, algo.designs = list(a2 = data.table(instance = "foo"))), "reserved keyword 'instance'")
+
+  prob.designs = c(prob.designs, list(p2 = data.table()))
+  ids = addExperiments(reg = reg, prob.designs = prob.designs, algo.designs = algo.designs, combine = "bind")
+  expect_data_table(ids, nrow = 7, key = "job.id")
+  expect_data_table(getJobPars(reg = reg), nrow = 16)
+
+  ids = addExperiments(reg = reg, prob.designs = prob.designs, algo.designs = algo.designs, combine = "crossprod")
+  expect_data_table(ids, nrow = 12, key = "job.id")
+  expect_data_table(getJobPars(reg = reg), nrow = 28)
 })
 
 if (FALSE) {
