+++ R/JobTables.R
@@ -133,8 +133,8 @@ getJobPars.ExperimentRegistry = function(ids = NULL, flatten = NULL, prefix = FA
   tab = mergedJobs(reg, ids, c("job.id", "pars", "problem", "algorithm"))
 
   if (flatten %??% qtestr(tab$pars, c("v1", "L"), depth = 2L)) {
-    new.cols = rbindlist(lapply(tab$pars, unlist, recursive = FALSE), fill = TRUE)
-    if (ncol(new.cols) > 0L) {
+    new.cols = rbindlist(.mapply(function(job.id, pars, ...) c(list(job.id = job.id), unlist(pars, recursive = FALSE)), tab, list()), fill = TRUE)
+    if (ncol(new.cols) >= 2L) {
       pattern = "^(prob|algo)\\.pars\\."
       replacement = if (prefix) "$1.par." else ""
       setnames(new.cols, names(new.cols), stri_replace_all_regex(names(new.cols), pattern, replacement))
