### 2016-11-25 d440b07
removed obsolete note
