### 2016-11-25 0341ea5
updated to new data.table
