### 2016-11-28 b7d1b2b
reduce overhead of doJobCollection
