### 2016-11-29 91dd9a6
improved parallelMap tests
