### 2016-11-29 f606b29
Update clusterFunctionsLSF.R
