### 2016-11-29 025ec5d
Update clusterFunctionsLSF.R
error message is "pending" rather than "unfinished" on my LSF 9.1.2.0
