### 2016-12-02 c3cc0c8
Update clusterFunctionsLSF.R (#45)
