### 2016-12-02 2c39d6a
Apply PR#45 also for OpenLava
