### 2016-12-05 #120 @arfon
@masalmon - would you be willing to review this submissions for JOSS?
