### 2016-12-06 627d4b2
cleanup ustamp/Sys.time calls
