### 2016-12-07 736f483
add mergeRegistries proto
