+++ tests/testthat/test_mergeRegistries.R
@@ -1,0 +1,25 @@
+context("mergeRegistries")
+
+test_that("mergeRegistries", {
+  target = makeRegistry(NA, make.default = FALSE)
+  f = function(.job, x) { if (x %in% c(2, 7)) file.create(file.path(.job$external.dir, "foo")); x^2 }
+  batchMap(f, 1:10, reg = target)
+
+  td = tempfile()
+  dir.create(td)
+  file.copy(target$file.dir, td, recursive = TRUE)
+  file.dir = file.path(td, basename(target$file.dir))
+  source = loadRegistry(file.dir, update.paths = TRUE)
+
+  submitJobs(data.table(job.id = 1:4, chunk = 1L), reg = target)
+  submitJobs(data.table(job.id = 6:9, chunk = c(1L, 1L, 1L, 2L)), reg = source)
+
+  mergeRegistries(source, target)
+  expect_data_table(findDone(reg = target), nrow = 8)
+
+  checkTables(target)
+
+  expect_set_equal(list.files(file.path(target$file.dir, "external")), as.character(c(2, 7)))
+  expect_equal(reduceResultsDataTable(reg = target)$V1, c(1,2,3,4,6,7,8,9)^2)
+  expect_file_exists(file.path(target$file.dir, "external", c("2", "7"), "foo"))
+})
