### 2016-12-07 9a2f9d9
updated NEWS [ci skip]
