+++ NEWS.md
@@ -3,6 +3,7 @@
 * `findExperiments` now has two additional arguments to match for regular expressions.
   The possibility to prefix a string with "~" to enable regular expression matching has been removed.
 * Fixed listing of jobs for `ClusterFunctionsLSF` and `ClusterFunctionsOpenLava` (thanks to @phaverty).
+* Fixed wrong key lookup in join functions
 
 # batchtools 0.9.0
 
