### 2016-12-07 76ec24e
pass updates table to post.submit hook
