### 2016-12-08 #135 @whedon
[REVIEW]: batchtools: Tools for R to work on batch systems
**Submitting author:** @mllg (<a href="http://orcid.org/0000-0001-9754-0393">Michel Lang</a>)
**Repository:** <a href="https://github.com/mllg/batchtools" target ="_blank">https://github.com/mllg/batchtools</a>
**Version:** v0.9.2
