### 2016-12-08 a3b8950
Updating paper metadata (#47)
