### 2016-12-08 663fdf9
fixed difftime bug
