+++ NAMESPACE
@@ -122,6 +122,7 @@ importFrom(brew,brew)
 importFrom(digest,digest)
 importFrom(progress,progress_bar)
 importFrom(rappdirs,user_config_dir)
+importFrom(stats,predict)
 importFrom(stats,runif)
 useDynLib(batchtools,count_not_missing)
 useDynLib(batchtools,fill_gaps)
