### 2016-12-08 5b942f1
changed col type to double; added update routine
