+++ R/resetJobs.R
@@ -19,7 +19,7 @@ resetJobs = function(ids = NULL, reg = getDefaultRegistry()) {
 
   info("Resetting %i jobs in DB ...", nrow(ids))
   cols = c("submitted", "started", "done", "error", "memory", "resource.id", "batch.id", "job.hash")
-  reg$status[ids, (cols) := list(NA_integer_, NA_integer_, NA_integer_, NA_character_, NA_real_, NA_integer_, NA_character_, NA_character_), on = "job.id"]
+  reg$status[ids, (cols) := list(NA_real_, NA_real_, NA_real_, NA_character_, NA_real_, NA_integer_, NA_character_, NA_character_), on = "job.id"]
 
   sweepRegistry(reg)
   invisible(ids)
