+++ man/lpt.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/estimateRuntimes.R
+\name{lpt}
+\alias{lpt}
+\title{Simple LPT Solver}
+\usage{
+lpt(x, ncpus = 1L)
+}
+\arguments{
+\item{x}{[\code{numeric}]\cr
+Numeric vector of elements to group.}
+
+\item{bins}{[\code{numeric(1)}]\cr
+Number of bins.}
+}
+\value{
+[\code{integer}] with values \dQuote{1} to \dQuote{bins} indicating bin membership.
+}
+\description{
+Maps numeric items in \code{x} into a fixed number of groups whose elements sum up to a similar value.
+A very simple greedy algorithm is used, which is not really optimized for speed.
+}
+\examples{
+x = 1:10
+bp = lpt(x, 3)
+xs = split(x, bp)
+print(xs)
+print(sapply(xs, sum))
+}
+\seealso{
+binpack estimateRuntimes
+}
+
