+++ NAMESPACE
@@ -30,6 +30,7 @@ export(ajoin)
 export(batchExport)
 export(batchMap)
 export(batchMapResults)
+export(binpack)
 export(btlapply)
 export(btmapply)
 export(cfBrewTemplate)
@@ -73,6 +74,7 @@ export(killJobs)
 export(ljoin)
 export(loadRegistry)
 export(loadResult)
+export(lpt)
 export(makeClusterFunctions)
 export(makeClusterFunctionsDocker)
 export(makeClusterFunctionsInteractive)
