### 2016-12-09 262b692
improve update routine
