+++ tests/testthat/test_submitJobs.R
@@ -10,7 +10,7 @@ test_that("submitJobs", {
 
   expect_integer(reg$status[1:2, resource.id], any.missing = FALSE)
   expect_character(reg$status[1:2, batch.id], any.missing = FALSE)
-  expect_integer(reg$status[1:2, submitted], any.missing = FALSE)
+  expect_numeric(reg$status[1:2, submitted], any.missing = FALSE)
   expect_true(is.na(reg$status[3, submitted]))
   x = reg$resources[1, resources][[1L]]
   y = insert(reg$default.resources, list(foo = "bar"))
