### 2016-12-09 f52afa7
updated NEWS [ci skip]
