### 2016-12-09 31e2f6d
Timestap accuracy (#48)
store timestamps with sub-seconds accuracy
