+++ man/Registry.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Registry.R
+% Please edit documentation in R/Registry.R, R/clearRegistry.R, R/sweepRegistry.R
 \name{Registry}
 \alias{Registry}
 \alias{clearRegistry}
@@ -26,11 +26,11 @@ loadRegistry(file.dir = getwd(), work.dir = NULL,
 
 saveRegistry(reg = getDefaultRegistry())
 
-sweepRegistry(reg = getDefaultRegistry())
+syncRegistry(reg = getDefaultRegistry())
 
 clearRegistry(reg = getDefaultRegistry())
 
-syncRegistry(reg = getDefaultRegistry())
+sweepRegistry(reg = getDefaultRegistry())
 }
 \arguments{
 \item{file.dir}{[\code{character(1)}]\cr
