+++ R/estimateRuntimes.R
@@ -23,6 +23,7 @@
 #'  \dQuote{observed} if runtime was observed).
 #'  The other element of the list named \dQuote{model}] contains the fitted random forest object.
 #' @export
+#' @seealso lpt binpack
 #' @examples
 #' # Create a simple toy registry
 #' tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)
@@ -69,14 +70,15 @@
 #' ids = est$runtimes[type == "estimated"]
 #' ids[, chunk := binpack(runtime, 3600)]
 #' print(ids)
-#' print(ids[, sum(runtime), by = chunk])
+#' print(ids[, list(runtime = sum(runtime)), by = chunk])
 #' \dontrun{
 #' submitJobs(ids, reg = tmp)
 #' }
 #'
-#'
-#' # Group jobs into a fixed number of chunks with similar runtime
-#'
+#' # Group jobs into 10 chunks with similar runtime
+#' ids = est$runtimes[type == "estimated"]
+#' ids[, chunk := lpt(runtime, 10)]
+#' print(ids[, list(runtime = sum(runtime)), by = chunk])
 estimateRuntimes = function(tab, ..., reg = getDefaultRegistry()) {
   assertRegistry(reg, sync = TRUE)
   data = copy(convertIds(reg, tab, keep.extra = names(tab)))
@@ -84,7 +86,6 @@ estimateRuntimes = function(tab, ..., reg = getDefaultRegistry()) {
   if (!requireNamespace("ranger", quietly = TRUE))
     stop("Please install package 'ranger' for runtime estimation")
 
-  runtime = NULL
   data[, "runtime" := as.numeric(getJobStatus(tab, reg)$time.running)]
   i = is.na(data$runtime)
   if (all(i))
@@ -128,87 +129,4 @@ print.RuntimeEstimate = function(x, n = 1L, ...) {
     catf("  Parallel : %s", ps(max(vnapply(split(rt, lpt(rt, n)), sum)), nc = nc))
   }
   catf("  Total    : %s", ps(total, nc = nc))
-}
-
-
-#' Simple Binpack Solver
-#'
-#' @description
-#' Maps numeric items in \code{x} into groups with sum less or equal than \code{capacity}.
-#' A very simple greedy algorithm is used, which is not really optimized for speed.
-#' If an element of \code{x} exceeds \code{capacity}, an error is thrown.
-#'
-#' @param x [\code{numeric}]\cr
-#'   Numeric vector of elements to group.
-#' @param capacity [\code{numeric(1)}]\cr
-#'   Maximum capacity of each bin, i.e., elements will be grouped so their sum does not exceed this limit.
-#' @return [\code{integer}] with values \dQuote{1} to \dQuote{n.bins} indicating bin membership.
-#' @export
-#' @seealso lpt estimateRuntimes
-#' @examples
-#' x = 1:10
-#' bp = binpack(x, 11)
-#' xs = split(x, bp)
-#' print(xs)
-#' print(sapply(xs, sum))
-binpack = function(x, capacity) {
-  assertNumeric(x, min.len = 1L, lower = 0, any.missing = FALSE, finite = TRUE)
-  assertNumber(capacity)
-
-  too.big = wf(x > capacity, use.names = FALSE)
-  if (length(too.big))
-    stopf("Capacity not sufficient. Item %i (x=%f) does not fit", too.big, x[too.big])
-
-  ord = order(x, decreasing = TRUE)
-  grp = integer(length(x))
-  sums = numeric(1L)
-  bin.count = 1L
-
-  for(j in ord) {
-    new.sums = sums + x[j]
[…]
