+++ tests/testthat/test_chunkIds.R
@@ -11,21 +11,21 @@ test_that("chunkIds", {
   ids = addExperiments(prob.designs, algo.designs, repls = repls, reg = reg)
   ids = getJobPars(reg = reg)[, c("job.id", "problem")]
 
-  expect_error(chunkIds(1:3, chunk.size = 1, n.chunks = 1, reg = reg), "either")
+  expect_error(suppressWarnings(chunkIds(1:3, chunk.size = 1, n.chunks = 1, reg = reg)), "either")
 
-  res = chunkIds(ids, n.chunks = 1, reg = reg)
+  res = suppressWarnings(chunkIds(ids, n.chunks = 1, reg = reg))
   expect_data_table(res, ncol = 2L, any.missing = FALSE, key = "job.id")
   expect_set_equal(names(res), c("job.id", "chunk"))
   expect_true(all(res$chunk == 1))
 
-  res = chunkIds(ids, n.chunks = 1, group.by = "problem", reg = reg)
+  res = suppressWarnings(chunkIds(ids, n.chunks = 1, group.by = "problem", reg = reg))
   expect_data_table(res, ncol = 2L, any.missing = FALSE, key = "job.id")
   expect_set_equal(names(res), c("job.id", "chunk"))
   res = ids[res]
   expect_true(all(res[problem == "p1", chunk] == 1))
   expect_true(all(res[problem == "p2", chunk] == 2))
 
-  res = chunkIds(ids, chunk.size = 10, group.by = "problem", reg = reg)
+  res = suppressWarnings(chunkIds(ids, chunk.size = 10, group.by = "problem", reg = reg))
   res = ids[res]
   tab = table(res$problem, res$chunk)
   expect_equal(as.numeric(tab[1, ]), rep(c(10, 0), c(3, 6)))
@@ -35,5 +35,5 @@ test_that("chunkIds", {
 test_that("chunkIds with defaults", {
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
   batchMap(identity, 1:10, reg = reg)
-  expect_error(chunkIds(reg = reg))
+  expect_error(suppressWarnings(chunkIds(reg = reg)))
 })
