### 2016-12-09 429bff1
added tests for chunk
