### 2016-12-09 b1f2835
c implementation for greedy schedulers
