+++ NAMESPACE
@@ -127,5 +127,7 @@ importFrom(progress,progress_bar)
 importFrom(rappdirs,user_config_dir)
 importFrom(stats,predict)
 importFrom(stats,runif)
+useDynLib(batchtools,c_binpack)
+useDynLib(batchtools,c_lpt)
 useDynLib(batchtools,count_not_missing)
 useDynLib(batchtools,fill_gaps)
