### 2016-12-11 mllg/batchtools#49 @masalmon
Links to other packages
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

The README states "As a successor of the packages BatchJobs and BatchExperiments, batchtools".

* Maybe you could explain why it is a successor of those packages, e.g. what it does better? Is it meant to replace the other two packages?

* In the README of those two packages I could not find a reference to `batchtools`. It might make sense to add a link to `batchtools` in the README (and even documentation for non Github users) of `BatchJobs` and `BatchExperiments`, if `batchtools` does some tasks better?
