### 2016-12-11 mllg/batchtools#49 @masalmon
I have to correct what I have just written, I see the packages are compared in [NEWS.md](https://github.com/mllg/batchtools/blob/master/NEWS.md), sorry. But my second suggestion remains valid I think.
