### 2016-12-11 mllg/batchtools#50 @masalmon
Link to the package website in the repository description?
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

The link to https://mllg.github.io/batchtools/ (really nice that the package has the website by the way) could be added to the description of the repository, so that it might be easier to see?
