### 2016-12-11 mllg/batchtools#51 @masalmon
Organization of the reference section of the website
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

In the [reference](https://mllg.github.io/batchtools/reference/index.html) section of the package website, each .Rd is an entry, and there are many of them. I wonder if it would make the documentation easier to read in that part if the author used a _pkgdown.yml for creating groups, see e.g. [this one](https://github.com/dirkschumacher/ompr/blob/master/_pkgdown.yml) with the resulting [reference section](https://dirkschumacher.github.io/ompr/reference/index.html)
