### 2016-12-11 mllg/batchtools#49 @masalmon
Another thought, the doc of `batchtools` (README, vignette?) could maybe include a mention of other packages in "Parallel computing: Resource managers and batch schedulers " from https://cran.r-project.org/web/views/HighPerformanceComputing.html (`batch` and `flowr`) and compare `batchtools` to them?
