### 2016-12-11 mllg/batchtools#52 @masalmon
Link to other vignette
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

In the Pi example vignette, there is "see one of the other vignettes for an example", maybe give the title of the mentioned vignette?
