### 2016-12-11 mllg/batchtools#51 @masalmon
The config file of pkgdown would also allow to change the order of the vignette, placing the Pi example on top (yep, this is a really minor comment :-) ).
