### 2016-12-11 mllg/batchtools#49 @masalmon
Similarly, in the Pi example vignette it is stated "The function batchMap(fun, ...) works analogously to Map(f, ...) of the base package.": maybe you could add a sentence about why to choose `batchMap` other `Map` (I guess the answer is that you can do more things with `batchMap`)?
