### 2016-12-11 mllg/batchtools#49 @masalmon
Is there any way to update the JSS paper to mention the new package? Or at least to add a warning in https://www.jstatsoft.org/article/view/v064i11 ?
