### 2016-12-11 mllg/batchtools#53 @masalmon
Appveyor badge
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

The Appveyor badge in the README is red and refers to the estimateRuntimes branch, while the build for master is not failing (but it is the one one gets when clicking on the badge). I'm not sure it's a real problem, I wonder if it would make more sense to have two badges or to show the Appveyor badge for the master branch only.
