### 2016-12-11 mllg/batchtools#54 @masalmon
JSS paper vs. JOSS paper
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

The JSS paper is mentioned as a resource, which is understandable. I have a few comments/questions on this:

* Is there any way to update the JSS paper to mention the new package? Or at least to add a warning in https://www.jstatsoft.org/article/view/v064i11 ?

* Could you precise more in the README what is still usable in the JSS paper (e.g. the section about " We use an ExperimentRegistry where the job definition is split into creating problems and algorithms. " according to one `batchtools` vignette) / what parts of the interfaces changed (besides having one single package now). I guess one can get this information by reading NEWS.md + the JSS paper, but NEWS.md contains other information too. This would also make the novelty of the software described in the JOSS paper clearer.

* The second paragraph of the abstract of the JSS paper is really nice (the list with letters), I wonder if it is allowed to have it in your README / in an intro vignette too.
