### 2016-12-11 mllg/batchtools#55 @masalmon
Flowcharts
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/135)

This is again only a suggestion: the JSS paper has a flowchart for `batchExperiments`, maybe a flowchart would make sense to explain the relationships between `batchtools` function?
