### 2016-12-11 mllg/batchtools#51 @masalmon
By the way I see the "See also" sections of the functions documentation is well used, so I guess adding groups in a config file for pkgdown wouldn't be too much work?
