### 2016-12-11 #135 @masalmon
Review

This is a really nice package! I think that I cannot realize how good it is because I've no parallel computing experience, but I can see it is a very complete package. @arfon by the way would it make sense to have a second review by someone a bit more experienced in the applications of such a package? Maybe the authors of the issues https://github.com/mllg/batchtools/issues/30 and https://github.com/mllg/batchtools/issues/44 ?

The software documentation / paper only needs minor revisions in my opinion. I have opened quite a few issues in the repo itself. Most of them are _suggestions_ regarding the organization of the documentation. My only real concern is the changes between the two packages as described in the JSS paper and the single package described in this JOSS submission: I think this deserves a bit more than pointing readers to the NEWS.md.

### Conflict of interest

- [ x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (such as being a major contributor to the software).
### General checks

- [ x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/mllg/batchtools">repository url</a>?
- [ x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
- [ x] **Version:** Does the release version given match the GitHub release (v0.9.0)?

Yes, but I wonder if v0.9.1 will the the version used for the JOSS paper in the end?
- [ x] **Authorship:** Has the submitting author (@mllg) made major contributions to the software?

### Functionality

- [x ] **Installation:** Does installation proceed as outlined in the documentation?
- [ x] **Functionality:** Have the functional claims of the software been confirmed?
- [ x] **Performance:** Have any performance claims of the software been confirmed?

### Documentation

- [x ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Just a small remark here: I guess you assume your users to know which e.g. scheduler to choose? Or would it make sense to point them to resources for choosing one?

- [x ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
- [ x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The reason I do not check this is because of the JSS paper remarks I mention at the beginning of this review.

- [x ] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

Except for the line " It would be great if you would help expand this collection to cover more exotic configurations. To do so, please send your template via mail or open a new pull request." in the Setup vignette, I did not see any contributions guideline.

### Software paper

Paper PDF: [10.21105.joss.00135.pdf](https://github.com/openjournals/joss-reviews/files/639511/10.21105.joss.00135.pdf)

- [ x] **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
- [ x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- [ x] **References:** Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?
