### 2016-12-11 bf41506
change appveyor badge to show status for master branch (fixes #53)
