+++ README.md
@@ -2,7 +2,7 @@
 
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/batchtools)](https://cran.r-project.org/package=batchtools)
 [![Build Status](https://travis-ci.org/mllg/batchtools.svg?branch=master)](https://travis-ci.org/mllg/batchtools)
-[![Build status](https://ci.appveyor.com/api/projects/status/ypp14tiiqfhnv92k?svg=true)](https://ci.appveyor.com/project/mllg/batchtools/branch/master)
+[![Build status](https://ci.appveyor.com/api/projects/status/1gdgk7twxrghi943/branch/master?svg=true)](https://ci.appveyor.com/project/mllg/batchtools-jgbhb/branch/master)
 [![Coverage Status](https://img.shields.io/coveralls/mllg/batchtools.svg)](https://coveralls.io/r/mllg/batchtools?branch=master)
 
 As a successor of the packages [BatchJobs](https://github.com/tudo-r/BatchJobs) and [BatchExperiments](https://github.com/tudo-r/Batchexperiments), batchtools provides a parallel implementation of Map for high performance computing systems managed by schedulers like Slurm, Torque, or SGE.
