### 2016-12-11 mllg/batchtools#53 @mllg
Good point! I changed the badge to show the status for the master branch.
