+++ man/ExperimentRegistry.Rd
@@ -104,7 +104,7 @@ ids = findExperiments(prob.pars = (n == 50), reg = tmp)
 getJobPars(ids, reg = tmp)
 
 # Submit jobs
-submitJobs(ids, reg = tmp)
+submitJobs(reg = tmp)
 waitForJobs(reg = tmp)
 
 # Reduce the results of algorithm a1
