### 2016-12-11 aad21a2
cleanup schedulers
