+++ src/lpt.c
@@ -1,17 +1,24 @@
 #include <R.h>
 #include <Rinternals.h>
+#define min(a, b) (((a) < (b)) ? (a) : (b))
 
 SEXP c_lpt(SEXP x_, SEXP order_, SEXP chunks_) {
     const double * x = REAL(x_);
-    const int * order = INTEGER(order_);
-    const int chunks = INTEGER(chunks_)[0];
     const R_len_t n = length(x_);
+    const int * order = INTEGER(order_);
+    const int chunks = min(INTEGER(chunks_)[0], n);
 
     SEXP res = PROTECT(allocVector(INTSXP, n));
     int * bin = INTEGER(res);
-    double * sums = calloc(chunks, sizeof(double));
+    double * sums = malloc(chunks * sizeof(double));
 
-    for (R_len_t i = 0; i < n; i++) {
+    for (R_len_t i = 0; i < chunks; i++) {
+        R_len_t ii = order[i] - 1;
+        bin[ii] = i + 1;
+        sums[i] = x[ii];
+    }
+
+    for (R_len_t i = chunks; i < n; i++) {
         R_len_t ii = order[i] - 1;
         R_len_t pos = 0;
         for (R_len_t j = 1; j < chunks; j++) {
