+++ src/binpack.c
@@ -3,36 +3,33 @@
 
 SEXP c_binpack(SEXP x_, SEXP order_, SEXP capacity_) {
     const double * x = REAL(x_);
+    const R_len_t n = length(x_);
     const int * order = INTEGER(order_);
     const double capacity = REAL(capacity_)[0];
-    const R_len_t n = length(x_);
+    R_len_t ii = order[0] - 1;
 
-    if (x[order[0] - 1] > capacity)
+    if (x[ii] > capacity)
         error("Capacity not sufficient. Largest item does not fit.");
 
     SEXP res = PROTECT(allocVector(INTSXP, n));
     int * bin = INTEGER(res);
-
     double * sums = malloc(n * sizeof(double));
-    R_len_t bins = 1;
+    R_len_t pos, bins = 1;
+    bin[ii] = pos + 1;
+    sums[0] += x[ii];
 
-    for (R_len_t i = 0; i < n; i++) {
-        R_len_t ii = order[i] - 1;
-        Rboolean assigned = FALSE;
-        for (R_len_t pos = 0; pos < bins; pos++) {
-            double tmp = sums[pos] + x[ii];
-            if (tmp <= capacity) {
-                bin[ii] = pos + 1;
-                sums[pos] = tmp;
-                assigned = TRUE;
+    for (R_len_t i = 1; i < n; i++) {
+        ii = order[i] - 1;
+
+        for (pos = 0; pos < bins + 1; pos++) {
+            if (sums[pos] + x[ii] <= capacity) {
                 break;
             }
         }
-        if (!assigned) {
-            bin[ii] = bins + 1;
-            sums[bins] = x[ii];
+        bin[ii] = pos + 1;
+        sums[pos] += x[ii];
+        if (pos > bins)
             bins++;
-        }
     }
 
     free(sums);
