+++ tests/testthat/test_chunk.R
@@ -21,6 +21,9 @@ test_that("chunk", {
   res = chunk(x, chunk.size = chunk.size)
   expect_integer(res, len = length(x), lower = 1, upper = length(x), any.missing = FALSE)
   expect_integer(table(res), lower = 1, upper = chunk.size, any.missing = FALSE)
+
+  expect_equal(chunk(numeric(0), chunk.size = 1), integer(0))
+  expect_equal(chunk(numeric(0), n.chunks = 1), integer(0))
 })
 
 test_that("binpack", {
@@ -43,6 +46,8 @@ test_that("binpack", {
   res = binpack(x, chunk.size = chunk.size)
   expect_integer(res, len = length(x), lower = 1, upper = length(x), any.missing = FALSE)
   expect_numeric(sapply(split(x, res), sum), lower = min(x), upper = chunk.size, any.missing = FALSE)
+
+  expect_equal(binpack(numeric(0), 1), integer(0))
 })
 
 test_that("lpt", {
@@ -51,7 +56,7 @@ test_that("lpt", {
   expect_integer(res, len = length(x), lower = 1, upper = n.chunks, any.missing = FALSE)
   expect_numeric(sapply(split(x, res), sum), len = min(length(x), n.chunks), lower = min(x), any.missing = FALSE)
 
-  x = runif(100); n.chunks = 2
+  x = runif(100); n.chunks = 3
   res = lpt(x, n.chunks)
   expect_integer(res, len = length(x), lower = 1, upper = n.chunks, any.missing = FALSE)
   expect_numeric(sapply(split(x, res), sum), len = min(length(x), n.chunks), lower = min(x), any.missing = FALSE)
@@ -65,4 +70,7 @@ test_that("lpt", {
   res = lpt(x, n.chunks)
   expect_integer(res, len = length(x), lower = 1, upper = n.chunks, any.missing = FALSE)
   expect_numeric(sapply(split(x, res), sum), len = min(length(x), n.chunks), lower = min(x), any.missing = FALSE)
+  expect_equal(unname(res), 10:1)
+
+  expect_equal(lpt(numeric(0), 1), integer(0))
 })
