### 2016-12-11 17ce925
added cross-vignette link (fixes #52)
