### 2016-12-11 mllg/batchtools#52 @mllg
Thanks! I've added a hyperlink.
