### 2016-12-11 a89103e
updated README: added related software part, as suggested in #49
