### 2016-12-11 mllg/batchtools#49 @mllg
I'm now mentioning the packages and the HPC task view in the readme. I can add some info to the README of BatchJobs/BatchExperiments, too. I'm unsure how to reach user not using GitHub.

@berndbischl Any thoughts on this?
