### 2016-12-12 mllg/batchtools#49 @masalmon
A not so nice way would be to have a message each time the users load BatchJobs or BadExperiments. And deprecation messages for some functions?
