### 2016-12-12 mllg/batchtools#49 @mllg
I think a call to `.Deprecated` would be too invasive because these messages are simple warnings. It would probably obscure the "regular" warnings raised if something really is not working like intended.

A packageStartupMessage sound like a good idea to me.
