### 2016-12-13 496344e
added missing wait in test for mergeRegistries
