### 2016-12-13 6609cdd
fixed test for mergeRegistries to respect user temp.dir
