+++ tests/testthat/test_waitForJobs.R
@@ -21,7 +21,7 @@ test_that("waitForJobs: detection of expired jobs", {
     submitJobs(ids, reg = reg)
     batch.ids = reg$status$batch.id
     reg$cluster.functions$killJob(reg, batch.ids[1])
-    expect_warning(waitForJobs(ids, reg = reg, sleep = 1))
+    expect_warning(waitForJobs(ids, reg = reg, sleep = 1), "disappeared")
   })
 })
 
