+++ tests/testthat/test_mergeRegistries.R
@@ -5,7 +5,7 @@ test_that("mergeRegistries", {
   f = function(.job, x) { if (x %in% c(2, 7)) file.create(file.path(.job$external.dir, "foo")); x^2 }
   batchMap(f, 1:10, reg = target)
 
-  td = tempfile()
+  td = file.path(target$temp.dir, basename(tempfile()))
   dir.create(td)
   file.copy(target$file.dir, td, recursive = TRUE)
   file.dir = file.path(td, basename(target$file.dir))
@@ -13,8 +13,11 @@ test_that("mergeRegistries", {
 
   submitAndWait(target, data.table(job.id = 1:4, chunk = 1L))
   submitAndWait(source, data.table(job.id = 6:9, chunk = c(1L, 1L, 1L, 2L)))
+  expect_data_table(findDone(reg = source), nrow = 4)
+  expect_data_table(findDone(reg = target), nrow = 4)
 
   mergeRegistries(source, target)
+  expect_data_table(findDone(reg = source), nrow = 4)
   expect_data_table(findDone(reg = target), nrow = 8)
 
   checkTables(target)
