### 2016-12-13 mllg/batchtools#54 @berndbischl
> Is there any way to update the JSS paper to mention the new package? Or at least to add a warning in https://www.jstatsoft.org/article/view/v064i11 ?

i really dont think that we can update the paper at JSS. but we can possibly put a warning message in batchjobs and batchexps so users are made aware of batchtools, that should really be sufficient?
