### 2016-12-13 mllg/batchtools#55 @berndbischl
can you please be more precise where you would expect such a flowchart to live?
(as the JOSS paper is basically an extended abstract)

do you mean to host this online, as part of the documentation?
