### 2016-12-13 mllg/batchtools#49 @berndbischl
> A not so nice way would be to have a message each time the users load BatchJobs or BadExperiments.

I think this is the best / most appropriate way to handle this. batchtools is a clear successor. there really is no downside to use the new package (except for minor name and API changes) , and all further development will happen in bt.

the only reason to still use BJ and BE is to "reproduce" your old code, at least IMHO, so living with a packageStartupMessage from now on seems really OK. and in all other cases users are made aware of the new better option
