### 2016-12-13 mllg/batchtools#49 @berndbischl
@mllg

> Maybe you could explain why it is a successor of those packages, e.g. what it does better? Is it meant to replace the other two packages?

but i think @masalmon is completely correct that it is very hard to see for new users *why* bt is better, in the README on GH and the DESC in the package. We really need to link to a little page that explains why bt is preferable (i.e. that it does not need a database and will there able to handle a) more jobs b) be a lot more robust on different systems)

what do you think?
