### 2016-12-13 mllg/batchtools#54 @masalmon
Yes I guess it'd be sufficient, after all the JSS readers will probably try to load the 2 packages at some point or google around a bit about updates to the packages.
