### 2016-12-13 mllg/batchtools#55 @masalmon
So this is really only a suggestion and it could be in one of the vignette for instance.
