### 2016-12-13 mllg/batchtools#55 @berndbischl
ok, thats what i wanted to here. i simply wanted to clarify that you are not talking about the paper itself
