### 2016-12-13 mllg/batchtools#55 @masalmon
ok sorry for being unclear :-)
