+++ _pkgdown.yml
@@ -1,9 +1,5 @@
 title: batchtools
 
-templates:
-  params:
-    bootswatch: readable
-
 reference:
   - title: Overview
     contents:
