### 2016-12-13 9dc0d55
prefix job hashes with 'job' (fixes #56)
