+++ R/JobCollection.R
@@ -49,7 +49,7 @@ createCollection = function(jobs, resources = list(), reg = getDefaultRegistry()
   jc$file.dir     = reg$file.dir
   jc$work.dir     = reg$work.dir
   jc$seed         = reg$seed
-  jc$job.hash     = digest(list(runif(1L), as.numeric(Sys.time())))
+  jc$job.hash     = stri_join("job", digest(list(runif(1L), as.numeric(Sys.time()))))
   jc$uri          = file.path(reg$file.dir, "jobs", sprintf("%s.rds", jc$job.hash))
   jc$log.file     = file.path(reg$file.dir, "logs", sprintf("%s.log", jc$job.hash))
   jc$packages     = reg$packages
