+++ NEWS.md
@@ -6,6 +6,7 @@
 * Fixed broken key lookup in some join functions.
 * Fixed a bug where `getJobTable()` returned `difftimes` with the wrong unit (e.g., in minutes instead of seconds).
 * Timestamps are now stored with sub-second accuracy.
+* Job hashes are now prefixed with literal string 'job' to ensure they start with a letter as required by some SGE systems.
 
 # batchtools 0.9.0
 
