### 2016-12-13 1eab63f
restructured docs
