+++ R/Problem.R
@@ -1,4 +1,4 @@
-#' @title Define Problems and Algorithms for Experiments
+#' @title Define Problems for Experiments
 #'
 #' @description
 #' Problems may consist of up to two parts: A static, immutable part (\code{data} in \code{addProblem})
@@ -6,29 +6,23 @@
 #' For example, for statistical learning problems a data frame would be the static problem part while
 #' a resampling function would be the stochastic part which creates problem instance.
 #' This instance is then typically passed to a learning algorithm like a wrapper around a statistical model
-#' (\code{fun} in \code{addAlgorithm}).
+#' (\code{fun} in \code{\link{addAlgorithm}}).
 #'
-#' The functions serialize all components to the file system and register the respective problem or algorithm
-#' names in the \code{\link{ExperimentRegistry}}.
+#' This function serialize all components to the file system and registers the problem in the \code{\link{ExperimentRegistry}}.
 #'
-#' \code{getProblemIds} and \code{getAlgorithmIds} can be used to retrieve the IDs of already defined problems
-#' and algorithms, respectively.
+#' \code{removeProblem} removes all jobs from the registry which depend on the specific problem.
+#' \code{getProblemIds} can be used to retrieve the IDs of already defined problems.
 #'
 #' @param name [\code{character(1)}]\cr
-#'   Unique identifier for the problem or algorithm.
+#'   Unique identifier for the problem.
 #' @param data [\code{ANY}]\cr
 #'   Static problem part. Default is \code{NULL}.
 #' @param fun [\code{function}]\cr
-#'   For \code{addProblem}, the function defining the stochastic problem part.
+#'   The function defining the stochastic problem part.
 #'   The static part is passed to this function with name \dQuote{data} and the \code{\link{Job}}/\code{\link{Experiment}}
 #'   is passed as \dQuote{job}.
 #'   Therefore, your function must have the formal arguments \dQuote{job} and \dQuote{data} (or dots \code{...}).
-#'
-#'   For \code{addAlgorithm}, the algorithm function. The static part is passed as \dQuote{data}, the generated
-#'   problem instance is passed as \dQuote{instance} and the \code{\link{Job}}/\code{\link{Experiment}} as \dQuote{job}.
-#'   Therefore, your function must have the formal arguments \dQuote{job}, \dQuote{data} and \dQuote{instance} (or dots \code{...}).
-#'
-#'   If you do not provide a function, it defaults to a function which just returns the data part (Problem) or the instance (Algorithm).
+#'   If you do not provide a function, it defaults to a function which just returns the data part.
 #' @param seed [\code{integer(1)}]\cr
 #'   Start seed for this problem. This allows the \dQuote{synchronization} of a stochastic
 #'   problem across algorithms, so that different algorithms are evaluated on the same stochastic instance.
@@ -42,9 +36,7 @@
 #'   different algorithms see different stochastic instances of the same problem.
 #' @template expreg
 #' @return [\code{Problem}]. Object of class \dQuote{Problem} (invisibly).
-#' @name ProblemAlgorithm
-#' @rdname ProblemAlgorithm
-#' @aliases Problem Algorithm
+#' @aliases Problem
 #' @export
 #' @examples
[…]

[…]
