### 2016-12-13 mllg/batchtools#49 @mllg
Here is the packageStartupMessage: https://github.com/tudo-r/BatchJobs/commit/d2789126f2d6660cb65dc78b7c7240072007aada

I'll refactor the NEWS and put an explanation as a new section into the README.
