+++ NEWS.md
@@ -10,30 +10,19 @@
 
 # batchtools 0.9.0
 
-The packages [BatchJobs](https://github.com/tudo-r/BatchJobs/) and [BatchExperiments](https://github.com/tudo-r/Batchexperiments) have been merged for the following reasons:
-* The packages are tightly connected which makes maintaining difficult, because the changes have to be synchronized and tested against the current CRAN versions for compatibility.
-* BatchExperiments violates CRAN policies by calling internal functions of BatchJobs.
-  The options to solve this are both subpar: (a) code duplication or (b) exporting many internal functions of BatchJobs.
-* Ongoing troubles with the data base and the file system.
-  Although we invested weeks to mitigate issues with locks of the SQLite data base or file system (staged queries, file system timeouts, ...), BatchJobs is working unreliable on some systems with a somewhat slow network file system.
+The packages [BatchJobs](https://github.com/tudo-r/BatchJobs/) and [BatchExperiments](https://github.com/tudo-r/Batchexperiments) have been merged into batchtools.
 
-[BatchJobs](https://github.com/tudo-r/BatchJobs/) and [BatchExperiments](https://github.com/tudo-r/Batchexperiments) will remain on CRAN.
-
-
-## Internal changes
+## Internal changes in comparison with BatchJobs
 
 * `batchtools` does not use SQLite anymore.
-  Instead, all the information is stored directly in the registry using [data.tables](https://cran.r-project.org/package=data.table) acting as an in-memory database.
+  Instead, all the information is stored directly in the registry using [data.tables](https://cran.r-project.org/package=data.table) acting as an in-memory database. As a side effect, many operations are much faster.
 * Nodes do not have to access the registry.
   [submitJobs()](https://mllg.github.io/batchtools/reference/submitJobs) stores a temporary object of type [JobCollection](https://mllg.github.io/batchtools/reference/JobCollection) on the file system which holds all the information necessary to execute a chunk of jobs via [doJobCollection()](https://mllg.github.io/batchtools/reference/doJobCollection) on the node.
   This avoids file system locks because each job accesses only one file exclusively.
 * `ClusterFunctionsMulticore` now uses the parallel package for multicore execution.
   `ClusterFunctionsSSH` can still be used to emulate a scheduler-like system which respects the work load on the local machine.
 
-
-## Interface changes
-
-### BatchJobs
+## User visible changes
 * `batchtools` remembers the last created or loaded Registry and sets it as default registry.
   This way, you do not need to pass the registry around anymore.
   If you need to work with multiple registries simultaneously on the other hand, you can still do so by explicitly passing registries to the functions.
@@ -45,16 +34,14 @@ The packages [BatchJobs](https://github.com/tudo-r/BatchJobs/) and [BatchExperim
       See the vignette on cluster functions for more information.
 * Most functions now return a [data.table](https://cran.r-project.org/package=data.table) which is keyed with the `job.id`.
   This way, return values can be joined together easily and efficient (see this [help page](https://mllg.github.io/batchtools/reference/JoinTables) for some examples).
-
-### BatchExperiments
 * The building blocks of a problem has been renamed from `static` and `dynamic` to the more intuitive `data` and `fun`.
   Thus, algorithm function should have the formal arguments `job`, `data` and `instance`.
 * The function `makeDesign` has been removed.
   Parameters can be defined by just passing a `data.frame` or `data.table` to [addExperiments](https://mllg.github.io/batchtools/reference/addExperiments).
   For exhaustive designs, use `expand.grid()` or `data.table::CJ()`.
 
-
 ## New features
+* Support for DockerSwarm via `ClusterFunctionsDocker`.
 * Jobs can now be tagged and untagged to provide an easy way to group them.
[…]
