### 2016-12-13 mllg/batchtools#49 @mllg
I've put a short section in the README of batchtools.

@berndbischl Can you add a link to the README of BJ/BE? I think we have done enough then to redirect users.
