### 2016-12-14 #135 @arfon
> This is a really nice package! I think that I cannot realize how good it is because I've no parallel computing experience, but I can see it is a very complete package. @arfon by the way would it make sense to have a second review by someone a bit more experienced in the applications of such a package? Maybe the authors of the issues mllg/batchtools#30 and mllg/batchtools#44 ?

Thanks for this review @masalmon. I agree it might make sense to solicit a second review here.

@HenrikBengtsson @jakob-r - as users of this package would either of you be willing to review this submission for JOSS? The process is very lightweight (basically going through the checklist above).
