### 2016-12-14 #135 @arfon
> @jakob-r has a conflict of interest. He is a coworker of mine.

Understood. Thanks for the heads up.
