### 2016-12-14 #135 @HenrikBengtsson
I'd be happy to review, but (regardless how easy it is) I will _not_ be able to get to this until January.

PS. I haven't started to use batchtools yet, but am definitely planning to dive into the details, cf. https://github.com/HenrikBengtsson/future.batchtools
