### 2016-12-14 e1128dd
cleanup binpack
