### 2016-12-15 48de42c
use -i flag for bsub/cfLSF
