### 2016-12-15 #135 @arfon
> I'd be happy to review, but (regardless how easy it is) I will not be able to get to this until January.

👍 thanks for the heads up @HenrikBengtsson - that should be fine.
