### 2016-12-15 dfd0811
fixed last commit for LSF
