### 2016-12-16 21334f5
print packageStartupMessage if debugme is enabled
