### 2016-12-16 04cdf22
add version in batchtools output
