### 2016-12-16 0054741
turn assertion into slave error in doJobCollection
