### 2016-12-19 16123f1
use stdin in cfLSF
