+++ man/runOSCommand.Rd
@@ -1,10 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clusterFunctions.R
+% Please edit documentation in R/runOSCommand.R
 \name{runOSCommand}
 \alias{runOSCommand}
 \title{Run OS Commands on Local or Remote Machines}
 \usage{
-runOSCommand(sys.cmd, sys.args = character(0L), nodename = "localhost")
+runOSCommand(sys.cmd, sys.args = character(0L), stdin = "",
+  nodename = "localhost")
 }
 \arguments{
 \item{sys.cmd}{[\code{character(1)}]\cr
@@ -12,6 +13,9 @@ Command to run.}
 
 \item{sys.args}{[\code{character}]\cr
 Arguments for \code{sys.cmd}.}
+
+\item{stdin}{[\code{character(1)}]\cr
+Argument passed to \code{\link[base]{system2}}.}
 
 \item{nodename}{[\code{character(1)}]\cr
 Name of the SSH node to run the command on. If set to \dQuote{localhost} (default), the command
