### 2016-12-19 118402e
added test for runOSCommand
