### 2016-12-20 286a41b
do not remove JobCollection for array jobs
