### 2016-12-20 2821a8a
cleanup old job files for array jobs
