### 2016-12-20 4a7c92a
worker/ssh cleanups
