### 2016-12-21 7e1390c
improve doJobCollection docs (fixes #61)
