+++ R/doJobCollection.R
@@ -11,6 +11,7 @@
 #' @param output [\code{character(1)}]\cr
 #'   Path to a file to write the output to. Defaults to \code{NULL} which means
 #'   that output is written to the active \code{\link[base]{sink}}.
+#'   Do not set this if your scheduler redirects output to a log file.
 #' @return [\code{character(1)}]: Hash of the \code{\link{JobCollection}} executed.
 #' @family JobCollection
 #' @export
