### 2016-12-22 7f0876b
depend on new progress package (#60)
