### 2016-12-22 373ab5e
handle results from previous submits (for #62)
