+++ R/submitJobs.R
@@ -194,6 +194,10 @@ submitJobs = function(ids = NULL, resources = list(), reg = getDefaultRegistry()
       }
     }
 
+    # remove old result files
+    fns = file.path(reg$file.dir, "results", sprintf("%s.rds", ids.chunk$job.id))
+    file.remove(fns[file.exists(fns)])
+
     repeat {
       runHook(reg, "pre.submit")
       now = ustamp()
