+++ man/reduceResults.Rd
@@ -39,6 +39,7 @@ Aggregated results in the same order as provided ids.
 \description{
 A version of \code{\link[base]{Reduce}} for \code{\link{Registry}} objects
 which iterates over finished jobs and aggregates them.
+All jobs must have terminated, an error is raised otherwise.
 }
 \note{
 If you have thousands of jobs, disabling the progress bar (\code{options(batchtools.progress = FALSE)})
