+++ NAMESPACE
@@ -29,6 +29,7 @@ export(ajoin)
 export(batchExport)
 export(batchMap)
 export(batchMapResults)
+export(batchReduce)
 export(btlapply)
 export(btmapply)
 export(cfBrewTemplate)
