+++ R/ExperimentRegistry.R
@@ -31,9 +31,8 @@
 #' ids = findExperiments(prob.pars = (n == 50), reg = tmp)
 #' getJobPars(ids, reg = tmp)
 #'
-#' # Chunk jobs per algorithm and submit them:
-#' ids = chunkIds(getJobPars(reg = tmp), group.by = "algorithm", reg = tmp)
-#' submitJobs(ids, reg = tmp)
+#' # Submit jobs
+#' submitJobs(reg = tmp)
 #' waitForJobs(reg = tmp)
 #'
 #' # Reduce the results of algorithm a1
