+++ R/zzz.R
@@ -24,8 +24,18 @@
 #' @importFrom brew brew
 #' @importFrom progress progress_bar
 #' @importFrom rappdirs user_config_dir
-#' @importFrom stats runif
+#' @importFrom stats runif predict
 "_PACKAGE"
+
+#' @title Deprecated function in the batchtools package
+#' @rdname batchtools-deprecated
+#' @name batchtools-deprecated
+#' @description
+#' The following functions have been deprecated:
+#' \tabular{rl}{
+#'   \code{chunkIds} \tab deprecated in favor of \code{\link{chunk}}, \code{\link{lpt}} and \code{\link{binpack}}\cr
+#' }
+NULL
 
 batchtools = new.env(parent = emptyenv())
 batchtools$debug = FALSE
