+++ NAMESPACE
@@ -16,6 +16,7 @@ S3method(print,ClusterFunctions)
 S3method(print,ExperimentRegistry)
 S3method(print,JobCollection)
 S3method(print,Registry)
+S3method(print,RuntimeEstimate)
 S3method(print,Status)
 S3method(print,SubmitJobResult)
 S3method(runHook,JobCollection)
@@ -29,15 +30,18 @@ export(ajoin)
 export(batchExport)
 export(batchMap)
 export(batchMapResults)
+export(binpack)
 export(btlapply)
 export(btmapply)
 export(cfBrewTemplate)
 export(cfHandleUnknownSubmitError)
 export(cfKillJob)
 export(cfReadBrewTemplate)
+export(chunk)
 export(chunkIds)
 export(clearRegistry)
 export(doJobCollection)
+export(estimateRuntimes)
 export(execJob)
 export(findDone)
 export(findErrors)
@@ -71,6 +75,7 @@ export(killJobs)
 export(ljoin)
 export(loadRegistry)
 export(loadResult)
+export(lpt)
 export(makeClusterFunctions)
 export(makeClusterFunctionsDocker)
 export(makeClusterFunctionsInteractive)
@@ -120,6 +125,9 @@ importFrom(brew,brew)
 importFrom(digest,digest)
 importFrom(progress,progress_bar)
 importFrom(rappdirs,user_config_dir)
+importFrom(stats,predict)
 importFrom(stats,runif)
+useDynLib(batchtools,c_binpack)
+useDynLib(batchtools,c_lpt)
 useDynLib(batchtools,count_not_missing)
 useDynLib(batchtools,fill_gaps)

[…]
