+++ tests/testthat/test_batchReduce.R
@@ -2,21 +2,26 @@ context("batchReduce")
 
 test_that("batchReduce", {
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
-  ids = batchReduce(function(aggr, x) aggr+x, 1:20, init=0, blocks = chunk(seq_along(xs), n.chunks = 10), reg = reg)
-  submitJobs(reg)
-  waitForJobs(reg)
-  y = reduceResults(reg, fun=function(aggr, job, res) aggr+res, init=0)
+  xs = 1:20
+  ids = batchReduce(function(aggr, x) aggr+x, xs, init = 0, blocks = chunk(seq_along(xs), n.chunks = 10), reg = reg)
+  expect_data_table(ids, nrow = 10, key = "job.id")
+
+  submitAndWait(reg = reg)
+  y = reduceResults(fun=function(aggr, job, res) aggr+res, init = 0, reg = reg)
   expect_equal(y, sum(1:20))
-  expect_equal(ids, 1:2)
+})
 
-  reg = makeTestRegistry()
-  ids = batchReduce(reg, function(aggr, x, y) aggr+x+y, 1:20, init=0, block.size=10,
-    more.args=list(y=1))
-  submitJobs(reg)
-  waitForJobs(reg)
-  y = reduceResults(reg, fun=function(aggr, job, res) aggr+res, init=0)
-  expect_equal(y, sum(1:20)+20)
 
-  reg = makeTestRegistry()
-  expect_equal(batchReduce(reg, function(aggr, x) aggr+x, integer(0L), init=0, block.size=10), integer(0L))
+test_that("batchReduce w/ more.args", {
+  reg = makeRegistry(file.dir = NA, make.default = FALSE)
+  xs = 1:20
+  blocks = sort(chunk(seq_along(xs), n.chunks = 10))
+  ids = batchReduce(function(aggr, x, y) aggr+x+y, 1:20, init = 100, blocks = blocks, more.args = list(y=1), reg = reg)
+  expect_data_table(ids, nrow = 10, key = "job.id")
+
+  submitAndWait(reg = reg)
+
+  expect_equal(as.integer(reduceResultsList(reg = reg)), viapply(split(xs, blocks), function(x) 100L + length(x) + sum(x), use.names = FALSE))
+  y = reduceResults(fun=function(aggr, job, res) aggr+res, init = 0, reg = reg)
+  expect_equal(y, sum(1:20) + 20 + uniqueN(blocks) * 100)
 })
