### 2017-01-02 mllg/batchtools#49 @mllg
I've added links to batchtools in the respective README and as packageStartupMessage.
