### 2017-01-02 mllg/batchtools#54 @mllg
I've ported the 2nd paragraph to the README.
