### 2017-01-02 3f2a804
refactored vignettes
