+++ R/JobCollection.R
@@ -36,9 +36,11 @@
 #' @examples
 #' tmp = makeRegistry(file.dir = NA, make.default = FALSE, packages = "methods")
 #' batchMap(identity, 1:5, reg = tmp)
-#' # resources are usually forwared from submitJobs()
+#'
+#' # resources are usually set in submitJobs()
 #' jc = makeJobCollection(1:3, resources = list(foo = "bar"), reg = tmp)
-#' print(ls.str(jc))
+#' ls(jc)
+#' jc$resources
 makeJobCollection = function(ids = NULL, resources = list(), reg = getDefaultRegistry()) {
   UseMethod("makeJobCollection", reg)
 }
