### 2017-01-02 mllg/batchtools#54 @mllg
The comparison/migration stuff is now in a own vignette rather than in the news. This hopefully makes it easier to find and compare.
