### 2017-01-02 mllg/batchtools#54 @masalmon
This looks nice @mllg !
