### 2017-01-02 c7de64d
news, readme (fixes #54)
