### 2017-01-02 mllg/batchtools#54 @mllg
I guess this is now done. Please re-open if I missed something.
