### 2017-01-02 mllg/batchtools#51 @mllg
@surmann added a `_pkgdown.yml` and the function reference is now grouped.

It is not perfect, but there are some technical issues which block further customization (https://github.com/hadley/pkgdown/issues/247, https://github.com/hadley/pkgdown/issues/248). I'll keep track of the changes in pkgdown and will eventually update the grouping.
