### 2017-01-02 mllg/batchtools#51 @masalmon
Ok. The current grouping alreadly looks good :+1:
