### 2017-01-02 mllg/batchtools#51 @mllg
Thanks! I consider this then fixed for now.
