### 2017-01-02 #135 @masalmon
I think most my comments about documentation have been taken into account now, great job. Only the contribution guidelines are missing or maybe I've missed something :slightly_smiling_face:

The only other remaining issue I opened is [this one](https://github.com/mllg/batchtools/issues/55) which was only a suggestion.
