### 2017-01-02 ac2c514
cleanup examples
