+++ R/batchMap.R
@@ -44,13 +44,7 @@
 #'
 #' # example for an expand.grid()-like operation on parameters
 #' tmp = makeRegistry(file.dir = NA, make.default = FALSE)
-#' ids = batchMap(paste, args = CJ(x = letters[1:3], y = 1:3), more.args = list(sep = ""), reg = tmp)
-#' getJobPars(reg = tmp)
-#' testJob(6, reg = tmp)
-#'
-#' # example on how to deal with bulky objects
-#' tmp = makeRegistry(file.dir = NA, make.default = FALSE)
-#' ids = batchMap(paste, args = CJ(x = letters[1:3], y = 1:3), more.args = list(sep = ""), reg = tmp)
+#' ids = batchMap(paste, args = CJ(x = letters[1:3], y = 1:3), reg = tmp)
 #' getJobPars(reg = tmp)
 #' testJob(6, reg = tmp)
 batchMap = function(fun, ..., args = list(), more.args = list(), reg = getDefaultRegistry()) {
