+++ man/batchMap.Rd
@@ -61,13 +61,7 @@ getJobPars(reg = tmp)
 
 # example for an expand.grid()-like operation on parameters
 tmp = makeRegistry(file.dir = NA, make.default = FALSE)
-ids = batchMap(paste, args = CJ(x = letters[1:3], y = 1:3), more.args = list(sep = ""), reg = tmp)
-getJobPars(reg = tmp)
-testJob(6, reg = tmp)
-
-# example on how to deal with bulky objects
-tmp = makeRegistry(file.dir = NA, make.default = FALSE)
-ids = batchMap(paste, args = CJ(x = letters[1:3], y = 1:3), more.args = list(sep = ""), reg = tmp)
+ids = batchMap(paste, args = CJ(x = letters[1:3], y = 1:3), reg = tmp)
 getJobPars(reg = tmp)
 testJob(6, reg = tmp)
 }
