### 2017-01-02 03bce9a
updated docs
