### 2017-01-03 0087465
added flowchart to vignette (fixes #55)
