### 2017-01-03 25f15c0
added contribution guidelines
