### 2017-01-03 #135 @mllg
I've fixed the last issue/suggestion and added contribution guidelines to the README. Waiting for @HenrikBengtsson now.

Thanks again @masalmon for your comments which substantially improved the package.
