### 2017-01-03 bd6799f
updated docs
