### 2017-01-08 90df94f
removed utf8 char
