+++ R/JobCollection.R
@@ -52,8 +52,8 @@ createCollection = function(jobs, resources = list(), reg = getDefaultRegistry()
   jc$work.dir     = reg$work.dir
   jc$seed         = reg$seed
   jc$job.hash     = stri_join("job", digest(list(runif(1L), as.numeric(Sys.time()))))
-  jc$uri          = file.path(reg$file.dir, "jobs", sprintf("%s.rds", jc$job.hash))
-  jc$log.file     = file.path(reg$file.dir, "logs", sprintf("%s.log", jc$job.hash))
+  jc$uri          = getJobFiles(reg$file.dir, jc$job.hash)
+  jc$log.file     = getLogFiles(reg$file.dir, jc$job.hash)
   jc$packages     = reg$packages
   jc$namespaces   = reg$namespaces
   jc$source       = reg$source
