+++ R/Logs.R
@@ -1,7 +1,7 @@
 #' @useDynLib batchtools fill_gaps
 readLog = function(id, missing.as.empty = FALSE, reg = getDefaultRegistry()) {
-  tab = reg$status[id, c("job.id", "done", "job.hash"), nomatch = NA]
-  log.file = file.path(reg$file.dir, "logs", sprintf("%s.log", tab$job.hash))
+  tab = reg$status[id, c("job.id", "job.hash"), nomatch = NA]
+  log.file = getLogFiles(reg$file.dir, tab$job.hash)
 
   if (is.na(tab$job.hash) || !file.exists(log.file)) {
     if (missing.as.empty)

[…]
