### 2017-01-10 f14ab47
remove obsolete result files after removeExperiments
