### 2017-01-16 834258c
small fixes
