### 2017-01-16 18c0620
renamed n.array.jobs -> array.jobs
