+++ R/resetJobs.R
@@ -18,8 +18,8 @@ resetJobs = function(ids = NULL, reg = getDefaultRegistry()) {
   ids = convertIds(reg, ids, default = noIds())
 
   info("Resetting %i jobs in DB ...", nrow(ids))
-  cols = c("submitted", "started", "done", "error", "memory", "resource.id", "batch.id", "job.hash")
-  reg$status[ids, (cols) := list(NA_real_, NA_real_, NA_real_, NA_character_, NA_real_, NA_integer_, NA_character_, NA_character_), on = "job.id"]
+  cols = c("submitted", "started", "done", "error", "memory", "resource.id", "batch.id", "array.id", "job.hash")
+  reg$status[ids, (cols) := list(NA_real_, NA_real_, NA_real_, NA_character_, NA_real_, NA_integer_, NA_character_, NA_integer_, NA_character_), on = "job.id"]
   fns = getResultFiles(reg$file.dir, ids$job.id)
   file.remove(fns[file.exists(fns)])
 
