+++ R/JobTables.R
@@ -29,6 +29,7 @@
 #'     \item{error}{Either \code{NA} if the job terminated successfully or the error message.}
 #'     \item{memory}{Estimate of the memory usage.}
 #'     \item{batch.id}{Batch ID as reported by the scheduler.}
+#'     \item{array.id}{Array ID as reported by the scheduler.}
 #'     \item{job.hash}{Unique string identifying the job or chunk.}
 #'     \item{time.queued}{Time in seconds (as \code{\link[base]{difftime}}) the job was queued.}
 #'     \item{time.running}{Time in seconds (as \code{\link[base]{difftime}}) the job was running.}
