+++ man/getJobTable.Rd
@@ -52,6 +52,7 @@ Registry. If not explicitly passed, uses the default registry (see \code{\link{s
     \item{error}{Either \code{NA} if the job terminated successfully or the error message.}
     \item{memory}{Estimate of the memory usage.}
     \item{batch.id}{Batch ID as reported by the scheduler.}
+    \item{array.id}{Array ID as reported by the scheduler.}
     \item{job.hash}{Unique string identifying the job or chunk.}
     \item{time.queued}{Time in seconds (as \code{\link[base]{difftime}}) the job was queued.}
     \item{time.running}{Time in seconds (as \code{\link[base]{difftime}}) the job was running.}
