+++ R/JobCollection.R
@@ -53,7 +53,7 @@ createCollection = function(jobs, resources = list(), reg = getDefaultRegistry()
   jc$seed         = reg$seed
   jc$job.hash     = stri_join("job", digest(list(runif(1L), as.numeric(Sys.time()))))
   jc$uri          = getJobFiles(reg$file.dir, jc$job.hash)
-  jc$log.file     = getLogFiles(reg$file.dir, jc$job.hash)
+  jc$log.file     = getLogFiles(reg$file.dir, jc$job.hash, array.id = NA_integer_)
   jc$packages     = reg$packages
   jc$namespaces   = reg$namespaces
   jc$source       = reg$source
