### 2017-01-16 bb3bdf5
fixes for slurm dortmund
