### 2017-01-16 4baa36c
fixes for slurm dortmund
