### 2017-01-16 ef20de5
Batch reduce (#65)
