### 2017-01-16 caf981f
use dash as array sep
