### 2017-01-17 b25417c
disable array jobs for sge
