### 2017-01-20 3d73038
more flexibility: store log.file instead of array.id
