### 2017-01-20 4a7cb22
more flexibility: store log.file instead of array.id
