### 2017-01-20 8d1f5b5
disabled array jobs for lsf
