+++ NEWS.md
@@ -1,5 +1,8 @@
 # batchtools 0.9.1
 
+* Full support for array jobs for Slurm and Torque.
+  Array jobs have been disabled for now for SGE and LSF (due to missing information about the output format) but will be re-enable in a future release.
+  Note that the variable `n.array.jobs` has been removed from `JobCollection` in favor of the new logical variable `array.jobs`.
 * New function `batchReduce()`.
 * `chunkIds()` has been deprecated. Use `chunk()`, `lpt()` or `binpack()` instead.
 * `findExperiments()` now has two additional arguments to match for regular expressions.
@@ -11,8 +14,6 @@
 * Deactivated swap allocation for `clusterFunctionsDocker`.
 * Timestamps are now stored with sub-second accuracy.
 * Job hashes are now prefixed with literal string 'job' to ensure they start with a letter as required by some SGE systems.
-* Better support for array jobs.
-  Note that `n.array.jobs` has been removed from `JobCollection` in favor of the logical `array.jobs`.
 
 # batchtools 0.9.0
 
