### 2017-01-20 5a6bbc0
speedup some tests
