### 2017-01-22 69f9f1b
cleanup tests
