+++ R/waitForJobs.R
@@ -22,7 +22,7 @@
 #' @export
 waitForJobs = function(ids = NULL, sleep = 10, timeout = 604800, stop.on.error = FALSE, reg = getDefaultRegistry()) {
   assertRegistry(reg, writeable = FALSE, sync = TRUE)
-  assertNumeric(sleep, len = 1L, lower = 0.2, finite = TRUE)
+  assertNumeric(sleep, len = 1L, lower = 0, finite = TRUE)
   assertNumeric(timeout, len = 1L, lower = sleep)
   assertFlag(stop.on.error)
   ids = convertIds(reg, ids, default = .findSubmitted(reg = reg))
