### 2017-01-27 mllg/batchtools#69 @HenrikBengtsson
Suggestion: Use special version for developing / "work in progress"
Triggered by my [JOSS review](https://github.com/openjournals/joss-reviews/issues/135#issuecomment-275495424):

What is your branch model and  for you Git / GitHub repository?:

* On CRAN you have batchtools 0.9.0.
* On GitHub you have tag `v0.9.0`.  I assume this corresponds exactly to that version on CRAN, correct?
* On GitHub `master` you have batchtools 0.9.1.  Is this your running development branch, i.e. an installed batchtools from this branch has version 0.9.1 regardless of when installed although the version may differ.  Correct?

If so, may I suggest to instead use `0.9.0-9000` for "work in progress"?  It would make it clear that it's work in progress (commonly used by others but also an eye-catching outlier version number).  Also, if one installing the master branch today and you later submit a fixed 0.9.1 version to CRAN, `update.packages()` will not pick up the new CRAN update and user will be stuck with an old developers version.

FYI, the `git flow` tool and workflow model (http://nvie.com/posts/a-successful-git-branching-model/) handles lots of the release / tag versions / develop / feature-branch needs very nicely.
