### 2017-01-27 mllg/batchtools#70 @HenrikBengtsson
https://mllg.github.io/batchtools/: AppVeyor link broken
AppVeyor link broken on https://mllg.github.io/batchtools/ is broken.
