### 2017-01-29 e9a2e36
updated docs (fixes #70)
