### 2017-01-29 mllg/batchtools#69 @mllg
> On GitHub you have tag v0.9.0. I assume this corresponds exactly to that version on CRAN, correct?

Correct.

> On GitHub master you have batchtools 0.9.1. Is this your running development branch, i.e. an installed batchtools from this branch has version 0.9.1 regardless of when installed although the version may differ. Correct?

Also correct.

> If so, may I suggest to instead use 0.9.0-9000 for "work in progress"? It would make it clear that it's work in progress (commonly used by others but also an eye-catching outlier version number). Also, if one installing the master branch today and you later submit a fixed 0.9.1 version to CRAN, update.packages() will not pick up the new CRAN update and user will be stuck with an old developers version.

Good point. I will set the devel version to 0.9.1-9000 now, next CRAN release will be 0.9.2.
