### 2017-01-29 400dc06
add column for registries created with previous versions
