+++ tests/testthat/test_getJobTable.R
@@ -6,12 +6,12 @@ test_that("getJobTable.Registry", {
   ids = batchMap(fun, i = 1:4, j = rep(1, 4), reg = reg)
 
   tab = getJobTable(reg = reg, flatten = FALSE)
-  expect_data_table(tab, nrows = 4, ncols = 13, key = "job.id")
+  expect_data_table(tab, nrows = 4, ncols = 14, key = "job.id")
   expect_list(tab$pars)
   expect_equal(tab$pars[[1]], list(i = 1L, j = 1))
 
   tab = getJobTable(reg = reg, flatten = TRUE)
-  expect_data_table(tab, nrows = 4, ncols = 13, key = "job.id")
+  expect_data_table(tab, nrows = 4, ncols = 14, key = "job.id")
   expect_null(tab[["pars"]])
   expect_equal(tab$i, 1:4)
   expect_equal(tab$j, rep(1, 4))
@@ -119,7 +119,7 @@ test_that("getJobTable.ExperimentRegistry", {
   ids = addExperiments(list(p1 = data.table(k = 1)), list(a1 = data.table(sq = 1:3)), reg = reg)
 
   tab = getJobTable(reg = reg, flatten = FALSE)
-  expect_data_table(tab, nrows = 3, ncols = 16, key = "job.id")
+  expect_data_table(tab, nrows = 3, ncols = 17, key = "job.id")
   expect_copied(tab, reg$status)
   expect_list(tab$pars)
   expect_equal(tab$pars[[1]], list(prob.pars = list(k = 1), algo.pars = list(sq = 1)))
@@ -129,7 +129,7 @@ test_that("getJobTable.ExperimentRegistry", {
   expect_equal(tab$algorithm[1], factor("a1"))
 
   tab = getJobTable(ids = 1:3, reg = reg, flatten = TRUE)
-  expect_data_table(tab, nrows = 3, ncols = 16, key = "job.id")
+  expect_data_table(tab, nrows = 3, ncols = 17, key = "job.id")
   expect_null(tab[["pars"]])
   expect_set_equal(tab$k, rep(1, 3))
   expect_set_equal(tab$sq, 1:3)
