### 2017-01-30 7da18f0
updated NEWS [ci skip]
