+++ NEWS.md
@@ -1,19 +1,19 @@
-# batchtools 0.9.1
+# batchtools 0.9.2
 
-* Full support for array jobs for Slurm and Torque.
-  Array jobs have been disabled for now for SGE and LSF (due to missing information about the output format) but will be re-enable in a future release.
+* Full support for array jobs on Slurm and Torque.
+  Array jobs have been disabled for SGE and LSF (due to missing information about the output format) but will be re-enable in a future release.
   Note that the variable `n.array.jobs` has been removed from `JobCollection` in favor of the new logical variable `array.jobs`.
-* New function `batchReduce()`.
 * `chunkIds()` has been deprecated. Use `chunk()`, `lpt()` or `binpack()` instead.
 * `findExperiments()` now has two additional arguments to match for regular expressions.
   The possibility to prefix a string with "~" to enable regular expression matching has been removed.
-* New function `estimateRuntimes()` which uses a random forest to predict runtimes of unfinished jobs.
+* New function `batchReduce()`.
+* New function `estimateRuntimes()`.
+* Timestamps are now stored with sub-second accuracy.
 * Fixed listing of jobs for `ClusterFunctionsLSF` and `ClusterFunctionsOpenLava` (thanks to @phaverty).
-* Fixed broken key lookup in some join functions.
+* Job hashes are now prefixed with the literal string 'job' to ensure they start with a letter as required by some SGE systems.
+* Fixed broken key lookup heuristic join functions.
 * Fixed a bug where `getJobTable()` returned `difftimes` with the wrong unit (e.g., in minutes instead of seconds).
 * Deactivated swap allocation for `clusterFunctionsDocker`.
-* Timestamps are now stored with sub-second accuracy.
-* Job hashes are now prefixed with literal string 'job' to ensure they start with a letter as required by some SGE systems.
 
 # batchtools 0.9.0
 
