### 2017-01-30 3bdebeb
ignore chunks.as.arrayjobs if not supported
