+++ R/waitForJobs.R
@@ -21,7 +21,6 @@
 #'   one job terminated with an exception.
 #' @export
 waitForJobs = function(ids = NULL, sleep = 10, timeout = 604800, stop.on.error = FALSE, reg = getDefaultRegistry()) {
-  "!DEBUG Starting waitForJobs"
   assertRegistry(reg, writeable = FALSE, sync = TRUE)
   assertNumeric(sleep, len = 1L, lower = 0, finite = TRUE)
   assertNumeric(timeout, len = 1L, lower = sleep)
@@ -43,7 +42,7 @@ waitForJobs = function(ids = NULL, sleep = 10, timeout = 604800, stop.on.error =
     return(TRUE)
 
   batch.ids = getBatchIds(reg)
-  "!DEBUG Using `nrow(ids)` ids and `nrow(batch.ids)` batch ids"
+  "!DEBUG `match.call()[1]`: Using `nrow(ids)` ids and `nrow(batch.ids)` batch ids"
   if (nrow(batch.ids) == 0L)
     return(nrow(.findErrors(reg, ids)) == 0L)
 
