### 2017-01-30 98d1fb4
Debug statements for waitForJobs() (#71)
