### 2017-01-30 8d312ce
prefixed debug statements with function names
