+++ R/syncRegistry.R
@@ -9,7 +9,7 @@
 #' @family Registry
 #' @export
 syncRegistry = function(reg = getDefaultRegistry()) {
-  "!DEBUG Triggered syncRegistry"
+  "!DEBUG syncRegistry: Triggered syncRegistry"
   fns = list.files(getUpdatePath(reg), full.names = TRUE)
   if (length(fns) == 0L)
     return(invisible(TRUE))
