### 2017-01-30 a085779
fix test for submitJobs
