### 2017-01-30 df9dcc8
fixed reduceResultsList/NULL
