### 2017-01-30 9a8dd4c
reworked sweepRegistry
