### 2017-01-30 ca36fe2
improve waitForJobs/debug messages
