### 2017-01-30 e6fa301
added more hooks
