+++ R/clusterFunctionsDocker.R
@@ -91,5 +91,5 @@ makeClusterFunctionsDocker = function(image, docker.args = character(0L), image.
   }
 
   makeClusterFunctions(name = "Docker", submitJob = submitJob, killJob = killJob, listJobsRunning = listJobsRunning,
-    store.job = TRUE, hooks = list(pre.submit = housekeeping, post.sync = housekeeping))
+    store.job = TRUE, hooks = list(pre.submit.job = housekeeping, post.sync = housekeeping))
 } # nocov end
