+++ R/clusterFunctionsLSF.R
@@ -20,10 +20,11 @@
 #'
 #' @templateVar cf.name lsf
 #' @template template
+#' @inheritParams makeClusterFunctions
 #' @return [\code{\link{ClusterFunctions}}].
 #' @family ClusterFunctions
 #' @export
-makeClusterFunctionsLSF = function(template = findTemplateFile("lsf")) { # nocov start
+makeClusterFunctionsLSF = function(template = findTemplateFile("lsf"), scheduler.delay = 1) { # nocov start
   template = cfReadBrewTemplate(template)
 
   # When LSB_BJOBS_CONSISTENT_EXIT_CODE = Y, the bjobs command exits with 0 only
@@ -72,5 +73,5 @@ makeClusterFunctionsLSF = function(template = findTemplateFile("lsf")) { # nocov
   }
 
   makeClusterFunctions(name = "LSF", submitJob = submitJob, killJob = killJob, listJobsQueued = listJobsQueued,
-    listJobsRunning = listJobsRunning, store.job = TRUE)
+    listJobsRunning = listJobsRunning, store.job = TRUE, scheduler.delay = scheduler.delay)
 } # nocov end
