+++ tests/testthat/test_ClusterFunctions.R
@@ -3,7 +3,8 @@ context("clusterFunctions")
 test_that("clusterFunctions constructor", {
   check = function(cf) {
     expect_is(cf, "ClusterFunctions")
-    expect_set_equal(names(cf), c("name", "submitJob", "killJob", "listJobsQueued", "listJobsRunning", "store.job", "array.var", "hooks"))
+    expect_set_equal(names(cf), c("name", "submitJob", "killJob", "listJobsQueued", "listJobsRunning",
+        "store.job", "array.var", "scheduler.delay", "hooks"))
     expect_output(print(cf), "ClusterFunctions for mode")
   }
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
