### 2017-01-31 b1ea3a4
fixed grepLogs for array jobs
