### 2017-01-31 3df4387
fixed empty log file test for array jobs
