### 2017-01-31 c1bfdfd
fixed default for grepLogs
