### 2017-01-31 307a779
added sleep to killJobs()
