+++ R/killJobs.R
@@ -50,7 +50,7 @@ killJobs = function(ids = NULL, reg = getDefaultRegistry()) {
   syncRegistry(reg = reg)
   cols = c("submitted", "started", "done", "error", "memory", "resource.id", "batch.id", "log.file", "job.hash")
   reg$status[tab[tab$killed], (cols) := list(NA_real_, NA_real_, NA_real_, NA_character_, NA_real_, NA_integer_, NA_character_, NA_character_, NA_character_)]
-
   saveRegistry(reg)
+  Sys.sleep(reg$cluster.functions$scheduler.delay)
   tab[, c("job.id", "batch.id", "killed")]
 }
