### 2017-01-31 d6e160d
hooks for killJobs
