+++ R/Hooks.R
@@ -15,6 +15,8 @@
 #'   \item{\code{post.do.collection}}{\code{function(reg, updates, cache, ...)}: Run after all jobs in the chunk are terminated on the slave.
 #'     \code{updates} is a \code{\link{data.table}} of updates which will be merged with the \code{\link{Registry}} by the master.
 #'     \code{cache} is an internal cache object.}
+#'   \item{\code{pre.kill}}{\code{function(reg, ids, ...)}: Run before any job is killed.}
+#'   \item{\code{post.kill}}{\code{function(reg, ids, ...)}: Run after jobs are killed. \code{ids} is the return value of \code{\link{killJobs}}.}
 #' }
 #'
 #' @param obj [\link{Registry} | \link{JobCollection}]\cr
