### 2017-01-31 21c0212
customizable sleep function (fixes #25)
