+++ R/waitForJobs.R
@@ -5,8 +5,9 @@
 #'
 #' @templateVar ids.default findSubmitted
 #' @template ids
-#' @param sleep [\code{numeric(1)}]\cr
-#'   Seconds to sleep between status updates. Default is \code{10}.
+#' @param sleep [\code{function(i)} | \code{numeric(1)}]\cr
+#'   Function which returns the duration to sleep in the \code{i}-th iteration.
+#'   Alternatively, you can pass a single positive numeric value.
 #' @param timeout [\code{numeric(1)}]\cr
 #'   After waiting \code{timeout} seconds, show a message and return
 #'   \code{FALSE}. This argument may be required on some systems where, e.g.,
@@ -20,11 +21,11 @@
 #'   successfully and \code{FALSE} if either the timeout is reached or at least
 #'   one job terminated with an exception.
 #' @export
-waitForJobs = function(ids = NULL, sleep = 10, timeout = 604800, stop.on.error = FALSE, reg = getDefaultRegistry()) {
+waitForJobs = function(ids = NULL, sleep = default.sleep, timeout = 604800, stop.on.error = FALSE, reg = getDefaultRegistry()) {
   assertRegistry(reg, writeable = FALSE, sync = TRUE)
-  assertNumeric(sleep, len = 1L, lower = 0, finite = TRUE)
-  assertNumeric(timeout, len = 1L, lower = sleep)
+  assertNumber(timeout, lower = 0)
   assertFlag(stop.on.error)
+  sleep = getSleepFunction(sleep)
   ids = convertIds(reg, ids, default = .findSubmitted(reg = reg))
 
   .findNotTerminated = function(reg, ids = NULL) {
@@ -47,6 +48,7 @@ waitForJobs = function(ids = NULL, sleep = 10, timeout = 604800, stop.on.error =
   timeout = Sys.time() + timeout
   ids.disappeared = noIds()
   pb = makeProgressBar(total = n.jobs, format = "Waiting (S::system R::running D::done E::error) [:bar] :percent eta: :eta")
+  i = 1L
 
   repeat {
     # case 1: all jobs terminated -> nothing on system
@@ -90,7 +92,8 @@ waitForJobs = function(ids = NULL, sleep = 10, timeout = 604800, stop.on.error =
     ids.disappeared = ids[!ids.on.sys, on = "job.id"]
     "!DEBUG [waitForJobs]: `nrow(ids.disappeared)` jobs disappeared"
 
-    Sys.sleep(sleep)
+    sleep(i)
+    i = 1 + 1L
     suppressMessages(syncRegistry(reg = reg))
     batch.ids = getBatchIds(reg)
     "!DEBUG [waitForJobs]: New batch.ids: `stri_flatten(batch.ids$batch.id, ',')`"
