+++ man/syncRegistry.Rd
@@ -23,4 +23,3 @@ Other Registry: \code{\link{clearRegistry}},
   \code{\link{loadRegistry}}, \code{\link{makeRegistry}},
   \code{\link{saveRegistry}}, \code{\link{sweepRegistry}}
 }
-
