+++ man/JobCollection.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/JobCollection.R
 \name{makeJobCollection}
-\alias{JobCollection}
 \alias{makeJobCollection}
+\alias{JobCollection}
 \title{JobCollection Constructor}
 \usage{
 makeJobCollection(ids = NULL, resources = list(),
@@ -59,4 +59,3 @@ jc$resources
 \seealso{
 Other JobCollection: \code{\link{doJobCollection}}
 }
-
