+++ man/cfKillJob.Rd
@@ -38,4 +38,3 @@ Other ClusterFunctionsHelper: \code{\link{cfBrewTemplate}},
   \code{\link{makeSubmitJobResult}},
   \code{\link{runOSCommand}}
 }
-
