+++ man/Worker.Rd
@@ -25,6 +25,7 @@ Worker
 
 \item{\code{status}}{Status of the worker; one of \dQuote{unknown}, \dQuote{available}, \dQuote{max.cpus} and \dQuote{max.load}.}
 }}
+
 \section{Methods}{
 
 \describe{
@@ -35,6 +36,7 @@ Worker
  \item{\code{kill(reg, batch.id)}}{Kill job matching the \dQuote{batch.id}.}
 }
 }
+
 \examples{
 \dontrun{
 # create a worker for the local machine and use 4 CPUs.
@@ -42,4 +44,3 @@ Worker$new("localhost", ncpus = 4)
 }
 }
 \keyword{datasets}
-
