+++ man/runHook.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/Hooks.R
 \name{runHook}
-\alias{Hook}
-\alias{Hooks}
 \alias{runHook}
+\alias{Hooks}
+\alias{Hook}
 \title{Trigger Evaluation of Custom Function}
 \usage{
 runHook(obj, hook, ...)
@@ -44,4 +44,3 @@ They can be specified via the \code{\link{ClusterFunctions}} and are trigged on 
   \item{\code{post.kill}}{\code{function(reg, ids, ...)}: Run after jobs are killed. \code{ids} is the return value of \code{\link{killJobs}}.}
 }
 }
-
