+++ man/makeClusterFunctionsSGE.Rd
@@ -58,4 +58,3 @@ Other ClusterFunctions: \code{\link{makeClusterFunctionsDocker}},
   \code{\link{makeClusterFunctionsTorque}},
   \code{\link{makeClusterFunctions}}
 }
-
