+++ man/makeClusterFunctionsDocker.Rd
@@ -27,11 +27,15 @@ Currently only user after \code{\link{submitJobs}}.}
 \description{
 Cluster functions for Docker/Docker Swarm (\url{https://docs.docker.com/swarm/}).
 
-The \code{submitJob} function executes \code{docker [docker.args] run --detach=true [image.args] [resources] [image] [cmd]}.
+The \code{submitJob} function executes
+\code{docker [docker.args] run --detach=true [image.args] [resources] [image] [cmd]}.
 Arguments \code{docker.args}, \code{image.args} and \code{image} can be set on construction.
-The \code{resources} part takes the named resources \code{ncpus} and \code{memory} from \code{\link{submitJobs}} and maps them to
-the arguments \code{--cpu-shares} and \code{--memory} (in Megabytes).
-To reliably identify jobs in the swarm, jobs are labeled with \dQuote{batchtools=[job.hash]} and named using the current login name and the job hash.
+The \code{resources} part takes the named resources \code{ncpus} and \code{memory}
+from \code{\link{submitJobs}} and maps them to the arguments \code{--cpu-shares} and \code{--memory}
+(in Megabytes). The resource \code{threads} is mapped to the environment variables \dQuote{OMP_NUM_THREADS}
+and \dQuote{OPENBLAS_NUM_THREADS}.
+To reliably identify jobs in the swarm, jobs are labeled with \dQuote{batchtools=[job.hash]} and named
+using the current login name (label \dQuote{user}) and the job hash (label \dQuote{batchtools}).
 
 \code{listJobsRunning} uses \code{docker [docker.args] ps --format=\{\{.ID\}\}} to filter for running jobs.
 
@@ -39,7 +43,7 @@ To reliably identify jobs in the swarm, jobs are labeled with \dQuote{batchtools
 
 These cluster functions use a \link{Hook} to remove finished jobs before a new submit and every time the \link{Registry}
 is synchronized (using \code{\link{syncRegistry}}).
-This is currently required because docker does not remove exited containers automatically.
+This is currently required because docker does not remove terminated containers automatically.
 Use \code{docker ps -a --filter 'label=batchtools' --filter 'status=exited'} to identify and remove terminated containers manually (or via a cron job).
 }
 \seealso{
