### 2017-02-01 32f09e9
hook cleanup
