+++ R/clusterFunctionsDocker.R
@@ -20,7 +20,8 @@
 #' These cluster functions use a \link{Hook} to remove finished jobs before a new submit and every time the \link{Registry}
 #' is synchronized (using \code{\link{syncRegistry}}).
 #' This is currently required because docker does not remove terminated containers automatically.
-#' Use \code{docker ps -a --filter 'label=batchtools' --filter 'status=exited'} to identify and remove terminated containers manually (or via a cron job).
+#' Use \code{docker ps -a --filter 'label=batchtools' --filter 'status=exited'} to identify and remove terminated
+#' containers manually (or usa a cron job).
 #'
 #' @param image [\code{character(1)}]\cr
 #'   Name of the docker image to run.
@@ -74,9 +75,23 @@ makeClusterFunctionsDocker = function(image, docker.args = character(0L), image.
     }
   }
 
+  listJobs = function(reg, filter = character(0L)) {
+    # use a workaround for DockerSwarm: docker ps does not list all jobs correctly, only
+    # docker inspect reports the status correctly
+    args = c(docker.args, "ps", "--format={{.ID}}", "--filter 'label=batchtools'", filter)
+    ids = runOSCommand("docker", args)
+    if (ids$exit.code != 0L)
+      stop("docker returned non-zero exit code")
+    if (length(ids$output) == 0L)
+      return(character(0L))
+
+    args = c(docker.args, "inspect", "--format '{{json .State.Status}}'", ids$output)
+    status = runOSCommand("docker", args)
+    ids$output[status$output == "\"running\""]
+  }
+
   housekeeping = function(reg, ...) {
-    args = c(docker.args, "ps", "-a", "--format={{.ID}}", "--filter 'label=batchtools'", "--filter 'status=exited'")
-    batch.ids = intersect(runOSCommand("docker", args)$output, reg$status$batch.id)
+    batch.ids = intersect(listJobs(reg, "--filter 'status=exited'"), reg$status$batch.id)
     if (length(batch.ids) > 0L)
       runOSCommand("docker", c(docker.args, "rm", batch.ids))
     invisible(TRUE)
@@ -90,11 +105,7 @@ makeClusterFunctionsDocker = function(image, docker.args = character(0L), image.
 
   listJobsRunning = function(reg) {
     assertRegistry(reg, writeable = FALSE)
-    args = c(docker.args, "ps", "--format={{.ID}}", "--filter 'label=batchtools'", sprintf("--filter 'user=%s'", user))
-    res = runOSCommand("docker", args)
-    if (res$exit.code == 0L)
-      return(res$output)
-    stop("docker returned non-zero exit code")
+    listJobs(reg, sprintf("--filter 'user=%s'", user))
   }
 
   makeClusterFunctions(name = "Docker", submitJob = submitJob, killJob = killJob, listJobsRunning = listJobsRunning,
