+++ NAMESPACE
@@ -126,6 +126,7 @@ importFrom(brew,brew)
 importFrom(digest,digest)
 importFrom(progress,progress_bar)
 importFrom(rappdirs,user_config_dir)
+importFrom(stats,pexp)
 importFrom(stats,predict)
 importFrom(stats,runif)
 useDynLib(batchtools,c_binpack)
