### 2017-02-02 81cbe1a
fix listJobs* for cfLSF
