### 2017-02-02 e13ec43
cleanup cfLSF
