### 2017-02-02 c70ad84
use faster string matching
