### 2017-02-02 mllg/batchtools#73 @HenrikBengtsson
NAMING CONVENTION: TORQUE not Torque
I'm pretty sure the "Terascale Open-source Resource and QUEue Manager" should be shortened using all upper-case letters as in "TORQUE".  See for instance,

* https://en.wikipedia.org/wiki/TORQUE
* http://www.adaptivecomputing.com/products/open-source/torque/

This affects batchtools as follows:
* Currently the documentation uses "Torque".
* Currently the API uses `makeClusterFunctionsTorque()`, which analogously to LSF and SGE should be renamed to `makeClusterFunctionsTORQUE()`.

See also Issue #28.
