### 2017-02-02 mllg/batchtools#74 @HenrikBengtsson
DOCUMENTATION: Typos / grammar mistakes
The documentation has several occurrences of

* "Univ**i**a Grid Engine" which should be "Univa Grid Engine" (http://www.univa.com/products/)

* "DockerSwarm" which should be "Docker Swarm" (https://www.docker.com/products/docker-swarm)
