### 2017-02-02 mllg/batchtools#74 @HenrikBengtsson
A few more mistakes in the docs:

* "All in all, the provided tools allow [<missing subject>] to work with many thousands or even millions of jobs in an organized and efficient manner."

* "maintenance" (not "maintaining") in "are tightly connected which makes maintaining difficult."

* "subtle" (not "subtile")
