### 2017-02-02 mllg/batchtools#75 @HenrikBengtsson
HELP: Rd mistakes
Rd typos:

* ?btlapply renders as "btlapply for lapply and btmapply for link[base]{mapply}."

* ?batchMap uses "[any] Arguments to vectorize over ..." whereas "[ANY]" is used elsewhere.
