### 2017-02-02 mllg/batchtools#76 @HenrikBengtsson
JOSS manuscript: Incorrect statement about NEWS
In https://github.com/openjournals/joss-reviews/files/639511/10.21105.joss.00135.pdf:

- "Important changes to its predecessors are summarized in the NEWS file to help users of BatchJobs/BatchExperiments migrating their cluster configuration and aid the transition to batchtools.": this seems to have moved to a vignette, so should probably be rephrased.

but the [`NEWS.md`](https://github.com/mllg/batchtools/blob/master/NEWS.md) file now refers to the vignette:

> batchtools 0.9.0
> Initial CRAN release. See this vignette for a brief comparison with BatchJobs/BatchExperiments.

The manuscript should be updated to refer to the vignette(s).

BTW / FYI, although CRAN renders `NEWS.md` on their package page, it is _not_ recognized by `utils::news()`, e.g.
```r
> news(package = "batchtools")
NULL
```
and the HTML help (as in `help.start()`) doesn't list any NEWS.  In other words, I think by using `NEWS.md` the NEWS information is only available on CRAN and GitHub and `file.show(system.file("NEWS.md", package = "batchtools"))`.   Should probably be reported / fixed in R itself, but I wanted to let you know.
