### 2017-02-02 mllg/batchtools#77 @HenrikBengtsson
DOCS: Define 'VPU'
The term 'VPU' doesn't seem to be defined anywhere and shouldn't be assumed user know what it is:
```sh
$ grep -F VPU -r
man/makeClusterFunctionsSocket.Rd:Number of VPUs of worker.
man/Worker.Rd:\item{\code{ncpus}}{Number of VPUs of worker. Set via constructor and defaults to the number of CPUs of the machine.}
man/makeClusterFunctionsMulticore.Rd:Number of VPUs of worker.
R/Worker.R:#' @field ncpus Number of VPUs of worker. Set via constructor and defaults to the number of CPUs of the machine.
man-roxygen/ncpus.R:#'   Number of VPUs of worker.
docs/reference/makeClusterFunctionsMulticore.html:Number of VPUs of worker.
docs/reference/makeClusterFunctionsSocket.html:Number of VPUs of worker.
docs/reference/Worker.html:    <p><dt><code>ncpus</code></dt><dd>Number of VPUs of worker. Set via constructor and defaults to the number of CPUs of the machine.</dd></p>
```

I'm sure you've already though about the potential confusion between argument name `ncpus` and "number of VPUs".
