### 2017-02-02 mllg/batchtools#78 @HenrikBengtsson
Porting to batchtools: BatchJobs::removeRegistry()?
What's the corresponding `batchtools` function to `BatchJobs::removeRegistry()`?  Something for cleaning out any permanent / intermediate files from the file system.

If my question make sense, I think this could be helpful to add to Section 'Porting to batchtools' in the 'Migration from BatchJobs/BatchExperiments' vignette.
