### 2017-02-02 mllg/batchtools#74 @HenrikBengtsson
A minor ambiguity and not very important, but "`ratio` % of" should probably be "`100 * ratio` % of" in:
```sh
$ grep -F "set for a split with" vignettes/v11_ExampleExperiment.Rmd
We define a problem function which returns the indices of the respective training and test set for a split with `ratio`% of the observations being in the test set:
```
or just "`ratio` _fraction_ of ...".
