### 2017-02-02 mllg/batchtools#79 @HenrikBengtsson
HELP: Where on the file system is saveRegistry() saving to?
`help("saveRegistry")` says:

> Stores the registry on the file system. [...]

But it doesn't mention where on the file system it saves the registry.  It also not clear to me how to programmatically find this information.
