### 2017-02-03 #135 @HenrikBengtsson
I've now completed my code / document review resulting in the following batchtools issues:

* https://github.com/mllg/batchtools/issues/69
* https://github.com/mllg/batchtools/issues/70
* https://github.com/mllg/batchtools/issues/73
* https://github.com/mllg/batchtools/issues/74
* https://github.com/mllg/batchtools/issues/75
* https://github.com/mllg/batchtools/issues/76
* https://github.com/mllg/batchtools/issues/77
* https://github.com/mllg/batchtools/issues/78
* https://github.com/mllg/batchtools/issues/79
* https://github.com/mllg/batchtools/issues/80
* https://github.com/mllg/batchtools/issues/81

Most of these are minor and not critical for accepting the manuscript.  The only one that could be consider **major** work / thought is [Issue 81](https://github.com/mllg/batchtools/issues/81) on 'RANDOMNESS: L'Cuyer RNG streams instead of ad-hoc random seeds?', which I think is an important topic.  I'd love to hear feedback from others who know / appreciate the problems of random number generation (and reproducibility) in parallel systems.

I've tested batchtools successfully on a local Linux machine, using the basic backends ("cluster functions") like "interactive" and "socket", which should work out of the box for everyone regardless of OS.   Now, I'm in the process of testing batchtools on a TORQUE / PBS system, on which I've previously used authors' BatchJobs package successfully / extensively.  I'm currently running into a few hiccups with batchtools ("Warning in waitForJobs() : Some jobs disappeared from the system") that I need to troubleshoot locally before reaching out / drawing any conclusions.  I would like to see a successful run on a real HPC cluster myself, although I'm in no doubt that the authors would post-publication make it work if it turns out that there is a bug.

Q to journal: "Authors: Does the paper.md file include a list of authors with their affiliations?" - I only see a link to the PDF, where's this Markdown file?  Also, in general, is there meant to be a separate issue / repos for discussing parts of a submitted paper, or is that meant to be done over at authors' software repos?
