+++ man/makeClusterFunctionsTORQUE.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clusterFunctionsTorque.R
-\name{makeClusterFunctionsTorque}
-\alias{makeClusterFunctionsTorque}
-\title{ClusterFunctions for OpenPBS/Torque Systems}
+% Please edit documentation in R/clusterFunctionsTORQUE.R
+\name{makeClusterFunctionsTORQUE}
+\alias{makeClusterFunctionsTORQUE}
+\title{ClusterFunctions for OpenPBS/TORQUE Systems}
 \usage{
-makeClusterFunctionsTorque(template = findTemplateFile("torque"),
+makeClusterFunctionsTORQUE(template = findTemplateFile("torque"),
   scheduler.delay = 1)
 }
 \arguments{
@@ -28,7 +28,7 @@ Currently only user after \code{\link{submitJobs}}.}
 [\code{\link{ClusterFunctions}}].
 }
 \description{
-Cluster functions for Torque/PBS (\url{http://www.adaptivecomputing.com/products/open-source/torque/}).
+Cluster functions for TORQUE/PBS (\url{http://www.adaptivecomputing.com/products/open-source/torque/}).
 
 Job files are created based on the brew template \code{template.file}. This file is processed
 with brew and then submitted to the queue using the \code{qsub} command. Jobs are killed using
