+++ DESCRIPTION
@@ -12,7 +12,7 @@ Description: As a successor of the packages 'BatchJobs' and 'BatchExperiments',
     (<http://www-03.ibm.com/systems/spectrum-computing/products/lsf/>),
     'OpenLava' (<http://www.openlava.org/>), 'Univia Grid Engine'/'Oracle Grid
     Engine' (<http://www.univa.com/>), 'Slurm' (<http://slurm.schedmd.com/>),
-    'Torque/PBS'
+    'TORQUE/PBS'
     (<http://www.adaptivecomputing.com/products/open-source/torque/>), or
     'Docker Swarm' (<https://docs.docker.com/swarm/>).
     A multicore and socket mode allow the parallelization on a local machines,
