+++ NAMESPACE
@@ -87,7 +87,7 @@ export(makeClusterFunctionsSGE)
 export(makeClusterFunctionsSSH)
 export(makeClusterFunctionsSlurm)
 export(makeClusterFunctionsSocket)
-export(makeClusterFunctionsTorque)
+export(makeClusterFunctionsTORQUE)
 export(makeExperimentRegistry)
 export(makeJob)
 export(makeJobCollection)
