### 2017-02-03 mllg/batchtools#73 @mllg
Fixed via 5b411a8ed45f63ddc66f933d1f175284685ef93a
