### 2017-02-03 a732b9f
fix typos (fixes #74)
