+++ DESCRIPTION
@@ -10,7 +10,7 @@ Description: As a successor of the packages 'BatchJobs' and 'BatchExperiments',
     this package provides a parallel implementation of the Map function for high
     performance computing systems managed by schedulers 'IBM Spectrum LSF'
     (<http://www-03.ibm.com/systems/spectrum-computing/products/lsf/>),
-    'OpenLava' (<http://www.openlava.org/>), 'Univia Grid Engine'/'Oracle Grid
+    'OpenLava' (<http://www.openlava.org/>), 'Univa Grid Engine'/'Oracle Grid
     Engine' (<http://www.univa.com/>), 'Slurm' (<http://slurm.schedmd.com/>),
     'TORQUE/PBS'
     (<http://www.adaptivecomputing.com/products/open-source/torque/>), or
