### 2017-02-03 cbd48bd
fix rd mistakes (fixes #75)
