### 2017-02-03 b6430e1
fix ref from NEWS to vignette in paper
