### 2017-02-03 mllg/batchtools#76 @mllg
I've updated the reference in a1c8bdd88949d3416226c16e829293b462301311. Will look into `news()` (which I never used before).
