### 2017-02-03 mllg/batchtools#81 @mllg
> I'm not claiming to be an expert in RNGs, but I believe these type of ad hoc set.seed() approaches is not considered correct by those who do know more about this. Another way to put it is, if it would be this simple, why do RNG methods like the L'Cuyer RNG streams even exist? So, my best guess is that the current approach is likely to produce suboptimal random numbers and in worst case highly correlated random numbers.

Not an RNG export, neither. As far as I understand it, the current approach picks different states from the mersenne twister which has `2^19937 - 1` states (`~4.3*10^6001`). I don't know how R translates the integer values to a RNG state, this is buried somewhere in the C code. Assuming that the integers lead to states that are evenly distributed in the period, you would need to draw `~2*10^5992 / jobs` numbers to get an overlap with the next stream. This is the best case analysis, of course. I guess L'Cuyer and other parallel RNGs exist so that you can get something which is guaranteed to not overlap, although it is very unlikely to happen if you use an ordinary RNG.

> I'd like to propose that you look at L'Cuyer RNG streams, which is implemented in the parallel package, meaning you don't need to add extra dependencies. The idea is that your seeds are generated as: [...]

I like the idea, that would probably be the right thing to implement. However, I'm worried about the runtime if you work on large experiments. I've tried to pre-compute the streams for 1e6 jobs (which is not unlikely to have if you use BatchExperiments) and stopped after some minutes. This would slow down `batchMap`/`addExperiments` significantly.
A more reasonable approach would be to ask the job to get the `i`-th stream where `i` depends on the job number. Unfortunately, you need to create all other preceding streams to get to the right one. So  job  number 1e6 would need to compute all streams which again would take far too long. I've also looked into `rlecluyer` and `rstreams`, but I was unable to find a way to get the `i`-th stream efficiently.
Suggestions welcome, of course. Maybe I've overlooked something important.

> If you're not comfortable of making use of L'Cuyer RNG streams right now, may I propose to allow for the option of the user to control this, e.g. seed = myNextSeed where myNextSeed can be a function.

This is something I'll look into. I guess this would be an extra argument for `makeRegistry`?
