### 2017-02-03 mllg/batchtools#81 @berndbischl
> If you're not comfortable of making use of L'Cuyer RNG streams right now, may I propose to allow for the option of the user to control this, e.g. seed = myNextSeed where myNextSeed can be a function.
> This is something I'll look into. I guess this would be an extra argument for makeRegistry?

Is this really reasonable? I don't think so. My arguments goes like this:

- Either our current approach is OK or it is not.
- If it is OK, that new option is not useful (at least now)
- It it isn't OK, who do you think will use this new option? Especially as we 3 guys (me included) don't know what is best...

My argument is: the RNG handling needs to be foolproof and correct in its default behavior. That's important. Imagine a user using our defaults, then figuring out that his results are flawed. Would you think an answer ala "hey you could have avoided this problem by using this pretty technical extra option..."
I would be quite furious.

@mllg I am happy to talk about this, probably on hangout is best. We could also ask Uwe for a further opinion.
