### 2017-02-04 #135 @arfon
@HenrikBengtsson - many thanks for the review. You've not updated the checklist at the top of this issue - is that because you don't have (GitHub) permissions to do this?

> Q to journal: "Authors: Does the paper.md file include a list of authors with their affiliations?" - I only see a link to the PDF, where's this Markdown file? Also, in general, is there meant to be a separate issue / repos for discussing parts of a submitted paper, or is that meant to be done over at authors' software repos?

The `paper.md` file is here: https://github.com/mllg/batchtools/blob/master/paper/paper.md . If there are issues with this then you should post them on the repository (https://github.com/mllg/batchtools) too.
