### 2017-02-04 #135 @HenrikBengtsson
@arfon, @masalmon cut'n'paste the checklist in a comment - shouldn't I do the same?  Who decides to check the top one? (given more than one reviewer?)
