### 2017-02-04 #135 @arfon
> @arfon, @masalmon cut'n'paste the checklist in a comment - shouldn't I do the same? Who decides to check the top one? (given more than one reviewer?)

Good point. I forgot there had been a review already 😄

I think it's fine for you to cut and paste in this case like @masalmon did already.
