### 2017-02-05 224e710
renamed 'VPU' to 'CPU' (fixes #77)
