+++ man/makeClusterFunctionsSocket.Rd
@@ -8,8 +8,8 @@ makeClusterFunctionsSocket(ncpus = NA_integer_)
 }
 \arguments{
 \item{ncpus}{[\code{integer(1)}]\cr
-Number of VPUs of worker.
-Default is to use all cores. The total number of cores "available" can be set via the option \code{mc.cores}
+Number of CPUs.
+Default is to use all logical cores. The total number of cores "available" can be set via the option \code{mc.cores}
 and defaults to the heuristic implemented in \code{\link[parallel]{detectCores}}.}
 }
 \value{
