### 2017-02-05 362face
added removeRegistry (fixes #78)
