+++ NAMESPACE
@@ -101,6 +101,7 @@ export(removeAlgorithms)
 export(removeExperiments)
 export(removeJobTags)
 export(removeProblems)
+export(removeRegistry)
 export(resetJobs)
 export(rjoin)
 export(runHook)
