### 2017-02-05 849bf16
added links from/to removeRegistry
