+++ man/sweepRegistry.Rd
@@ -19,5 +19,6 @@ and data base informations.
 Other Registry: \code{\link{clearRegistry}},
   \code{\link{getDefaultRegistry}},
   \code{\link{loadRegistry}}, \code{\link{makeRegistry}},
-  \code{\link{saveRegistry}}, \code{\link{syncRegistry}}
+  \code{\link{removeRegistry}}, \code{\link{saveRegistry}},
+  \code{\link{syncRegistry}}
 }
