### 2017-02-05 mllg/batchtools#78 @mllg
I simply forgot about this one. Now in batchtools, with a mention to `sweepRegistry`.
