### 2017-02-05 0f0e098
add info about destination of saveRegistry (fixes #79)
