### 2017-02-05 mllg/batchtools#80 @mllg
Jobs which fail to find a JobCollection file will terminate with an error but appear as "expired" on the master because they were unable to communicate anything back. The log files should be informative if the template file redirects the output and does not rely on `batchtools` to do so.

However, there are several mechanism which should guard against this:

1. `saveRDS` is blocking the session and `submitJob` is called after `saveRDS`.
2. I use this little helper [here](https://github.com/mllg/batchtools/blob/master/R/helpers.R#L40) to write the files and then wait for the file system / file system caches to find the file.
3. The write/rename approach is already implemented for registries where I want to be extra fail-safe and also keep a backup. I can port this to other write ops, too. The only possible downside I see is that this inflicts many more stat system calls. I'll investigate if this is a problem on some network file systems.
