### 2017-02-05 652e07f
always use safe write ops
