### 2017-02-05 8889d10
fixed removeRegistry test
