+++ tests/testthat/test_removeRegistry.R
@@ -3,7 +3,7 @@ context("removeRegistry")
 test_that("removeRegistry", {
   reg = makeRegistry(file.dir = NA, make.default = FALSE)
   expect_true(dir.exists(reg$file.dir))
-  expect_message({res <- removeRegistry(0, reg = reg)}, "Recursively removing")
+  res = removeRegistry(0, reg = reg)
   expect_true(res)
   expect_false(dir.exists(reg$file.dir))
 })
