+++ .travis.yml
@@ -1,2 +1,3 @@
 dist: trusty
 sudo: false
+language: r
