### 2017-02-05 97ac05c
re-enabled mpi test if Rmpi is installed
