### 2017-02-05 101f552
Switch to container build (#84)
