+++ .travis.yml
@@ -12,4 +12,4 @@ r_packages:
   - covr
 
 after_success:
-  - if [[ "${TRAVIS_R_VERSION}" == "stable" ]]; then Rscript -e 'covr::coveralls()'; fi
+  - if [[ "${TRAVIS_R_VERSION_STRING}" == "release" ]]; then Rscript -e 'covr::coveralls()'; fi
