### 2017-02-06 mllg/batchtools#85 @HenrikBengtsson
BUG: Post waitForJobs(), reduceResultsList() returns list with `NULL`:s and a bit later the values
# Issue

Using `makeClusterFunctionsTorque()` on a TORQUE compute cluster, at first `reduceResultsList()` returns list of `NULL` elements but a bit later a list of the actual values.  This even after `waitForJobs()` returns `TRUE`.  I suspect this is due to the infamous NFS delay and not polling the existence of the result files.

## Example
Sourcing the following `test.R` script:
```r
library("batchtools")
unlink("registry", recursive = TRUE)
stopifnot(!file_test("-d", "registry"))
reg <- makeRegistry("registry")
cat("#!/bin/bash\nRscript -e 'batchtools::doJobCollection(\"<%= uri %>\")'\n", file = "my.tmpl")
reg$cluster.functions <- makeClusterFunctionsTorque(template = "my.tmpl")
print(reg)

jobs <- batchMap(fun = identity, 1:3, reg = reg)
print(jobs)

submitJobs(reg = reg)

done <- waitForJobs(reg = reg)
stopifnot(done)

t0 <- Sys.time()
count <- 1L
repeat {
  y <- reduceResultsList(reg = reg)
  str(list(count = count, dt = Sys.time() - t0, y = y))
  if (!is.null(y[[1]])) break
  Sys.sleep(0.1)
  count <- count + 1L
}
```

gives the following:
```sh
> library("batchtools")
> unlink("registry", recursive = TRUE)
> stopifnot(!file_test("-d", "registry"))
> reg <- makeRegistry("registry")
> cat("#!/bin/bash\nRscript -e 'batchtools::doJobCollection(\"<%= uri %>\")'\n", file = "my.tmpl")
> reg$cluster.functions <- makeClusterFunctionsTorque(template = "my.tmpl")
> print(reg)
Job Registry
  Name    : Torque
  File dir: /home/henrik/batchtools-review/registry
  Work dir: /home/henrik/batchtools-review
  Jobs    : 0
  Seed    : 567656413

> jobs <- batchMap(fun = identity, 1:3, reg = reg)
Adding 3 jobs ...

> print(jobs)
   job.id
1:      1
2:      2
3:      3

> submitJobs(reg = reg)
Submitting 3 jobs in 3 chunks using cluster functions 'Torque' ...

> done <- waitForJobs(reg = reg)
> stopifnot(done)

> t0 <- Sys.time()
> count <- 1L
> repeat {
+   y <- reduceResultsList(reg = reg)
+   str(list(count = count, dt = Sys.time() - t0, y = y))
+   if (!is.null(y[[1]])) break
+   count < .... [TRUNCATED]
List of 3
 $ count: int 1
 $ dt   :Class 'difftime'  atomic [1:1] 0.00505
  .. ..- attr(*, "units")= chr "secs"
 $ y    :List of 3
  ..$ : NULL
  ..$ : NULL
  ..$ : NULL
List of 3
 $ count: int 2
 $ dt   :Class 'difftime'  atomic [1:1] 0.017
  .. ..- attr(*, "units")= chr "secs"
 $ y    :List of 3
  ..$ : NULL
  ..$ : NULL
  ..$ : NULL
[...]
List of 3
 $ count: int 425
 $ dt   :Class 'difftime'  atomic [1:1] 4.87
  .. ..- attr(*, "units")= chr "secs"
 $ y    :List of 3
  ..$ : NULL
  ..$ : NULL
  ..$ : NULL
List of 3
 $ count: int 426
 $ dt   :Class 'difftime'  atomic [1:1] 4.89
  .. ..- attr(*, "units")= chr "secs"
 $ y    :List of 3
  ..$ : int 1
  ..$ : int 2
  ..$ : int 3
>
```

# Troubleshooting / Suggestion

It looks like `batchtools:::.reduceResultsList()`, which is called by `reduceResultsList()`, [silently skips reading any result files for which `file.exists(fns)` returns `FALSE`](https://github.com/mllg/batchtools/blob/master/R/reduceResults.R#L158-L163).

If that's indeed the intended behavior, then I cannot tell whether the bug is in `reduceResultsList()` assuming all the results files should be there when `batchtools:::.reduceResultsList()` is called or in `waitForJobs()` that should not return until all the result files are there.

As my example code shows, I suspect that there's a delay in the network file system (NFS) causing the already written result files to not be visible from the master machine until several seconds later.  The `file.exists()` suggests this.

FYI, it looks like calling `dir(path = file.path(reg$file.path, "results"))` forces NFS to sync its view such that `file.exists()` returns `TRUE`.  However, I don't know whether that is a bullet proof solution.

To me it seems like batchtools (`waitForJobs()` or something) needs to pool for the result files before trying they are queried.

LATE UPDATES:
* I see that `batchtools:::.loadResult()` and hence `loadResult()` gives an error if the file is not there.
* I see from `help("reduceResultsList") that the "otherwise `NULL` behavior" is expected from that function.
* I see that `reduceResults()` will try to read the result files directly, so if they're not there an error will be generated.  PS / feedback. The different default / error behavior to `reduceResultsList()` is a bit confusing given their similarity in names.

# Session information

This is on a Scyld cluster with TORQUE / PBS + Moab.
```r
> sessionInfo()
R version 3.3.2 (2016-10-31)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Red Hat Enterprise Linux Server release 6.6 (Santiago)

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] batchtools_0.9.1-9000 data.table_1.10.4     R.archive_0.3.0

 [1] prettyunits_1.0.2 digest_0.6.12     assertthat_0.1    R.methodsS3_1.7.1
 [5] rappdirs_0.3.1    R6_2.2.0          backports_1.0.5   magrittr_1.5
 [9] progress_1.1.2    stringi_1.1.2     R.oo_1.21.0       R.utils_2.5.0
[13] brew_1.0-6        checkmate_1.8.2   tools_3.3.2
```
