### 2017-02-07 mllg/batchtools#85 @HenrikBengtsson
Using:
```r
file_exists_1 <- function(x) { isdir <- file.info(x)$isdir; !is.na(isdir) & !isdir }
file_exists_2 <-  function(x) { basename(x) %in% list.files(dirname(x)) }
```
on the first result file:
```r
pathname <- file.path(reg$file.dir, "results", "1.rds")
```
I get:
```sh
List of 6
 $ count        : int 1
 $ file.exists  : logi FALSE
 $ file_exists_1: logi FALSE
 $ file_exists_2: logi TRUE
 $ dt           :Class 'difftime'  atomic [1:1] 0.0138
  .. ..- attr(*, "units")= chr "secs"
 $ y            :List of 3
  ..$ : NULL
  ..$ : NULL
  ..$ : NULL
List of 6
 $ count        : int 2
 $ file.exists  : logi TRUE
 $ file_exists_1: logi TRUE
 $ file_exists_2: logi TRUE
 $ dt           :Class 'difftime'  atomic [1:1] 0.165
  .. ..- attr(*, "units")= chr "secs"
 $ y            :List of 3
  ..$ : int 1
  ..$ : int 2
  ..$ : int 3
```

I suspect that the first call to `file_exists_2(pathname)` triggers the NFS cache to be updated.  This is in line with my previous comment where I noticed that `dir(path = file.path(reg$file.path, "results"))` seems fix it.
