### 2017-02-07 c93d1ce
renamed scheduler.delay; introduced fs.latency
