### 2017-02-07 mllg/batchtools#85 @mllg
I've started with a simple `waitForFiles()` function [here](https://github.com/mllg/batchtools/pull/86).
You can set `fs.latency` in `makeClusterFunctions*` to a positive number (e.g., 30) to let `waitForJobs()` also wait for the results.

If the approach works, I will try to generalize and apply the same approach to other file system ops.
