### 2017-02-08 4791e87
consistent handling of missing results
