### 2017-02-10 mllg/batchtools#85 @HenrikBengtsson
Would it make sense to make `fs.latency=30` (the default) rather than zero?  This would cover more users (one less thing to worry about).  The downside is that users might not be / become aware of their NFS cache delay, but not sure if that's a big problem (e.g. I doubt they would go an optimize the settings to lower it anyway).
