### 2017-02-10 mllg/batchtools#85 @HenrikBengtsson
BTW, I wonder if there's could be a well defined / documented system call "out there" that triggers a blocking updating of the NFS cache?  Maybe `list.files() / dir()` calls this, which is why it appears to force NFS to be updated (at least on my system).
