### 2017-02-11 mllg/batchtools#87 @HenrikBengtsson
BUG: Variables starting with a period are exported but not imported
Variables starting with a period are exported but not imported.

# Example
```r
library("batchtools")
reg <- makeRegistry(NA)
objs <- list(a = 1, b = 2, .c = 3, d = 4)
str(objs)
batchExport(export = objs, reg = reg)
dir(file.path(reg$file.dir, "exports"), all.files = TRUE, full.names = TRUE)

batchMap(fun = function(i) {
    list(i = i, a = a, b = b, .c = .c, d = d)
}, i = 1L, reg = reg)

submitJobs(reg = reg)

ans <- waitForJobs(reg = reg)
stopifnot(isTRUE(ans))

y <- reduceResultsList(reg = reg)
str(y)
```
results in:
```r
❯ getErrorMessages(reg = reg)
   job.id terminated error                                         message
1:      1       TRUE  TRUE Error in (function (i)  : object '.c' not found
```

# Troubleshooting
This seems to be because [`loadRegistryDependencies()`](https://github.com/mllg/batchtools/blob/master/R/Registry.R#L247-L248) does not use `all.names = TRUE` in:
```r
  path = file.path(x$file.dir, "exports")
  fns = list.files(path, pattern = "\\.rds$")
```

# See also
BatchJobs had the exact same problem; https://github.com/tudo-r/BatchJobs/issues/103
