### 2017-02-11 mllg/batchtools#85 @HenrikBengtsson
Just reporting back. Trying which what's on the master branch right now, I get:
```r
> done <- waitForJobs(reg = reg)
Syncing 3 files ...
> done
[1] TRUE
> y <- reduceResultsList(reg = reg)
Error in gzfile(file, "rb") : cannot open the connection
In addition: Warning message:
In gzfile(file, "rb") :
  cannot open compressed file '/home/henrik/registry/results/1.rds', probable reason 'No such file or directory'
```
which I consider an improvement, because it's better to get an error than an incorrect value (= NULL).

And, sure, enough, I can refresh the NFS cache by calling:
```r
> dir(file.path(reg$file.dir, "results"))
[1] "1.rds" "2.rds" "3.rds"
```
and after this, I can read the results:
```r
> y <- reduceResultsList(reg = reg)
> str(y)
List of 3
 $ : num 1
 $ : num 2
 $ : num 3
```
