### 2017-02-13 mllg/batchtools#76 @mllg
Talked to Uwe about this. We agreed that this is really something to be fixed by R core. He'll look into it.
