### 2017-02-13 b7dcdd0
improve waitForFiles; default timeout is 65
