### 2017-02-13 46e2bfe
Handle FS latency (#86)
