+++ man/makeClusterFunctions.Rd
@@ -7,7 +7,7 @@
 \usage{
 makeClusterFunctions(name, submitJob, killJob = NULL, listJobsQueued = NULL,
   listJobsRunning = NULL, array.var = NA_character_, store.job = FALSE,
-  scheduler.delay = 0, hooks = list())
+  scheduler.latency = 0, fs.latency = NA_real_, hooks = list())
 }
 \arguments{
 \item{name}{[\code{character(1)}]\cr
@@ -46,9 +46,15 @@ Default is \code{NA} for no array support.}
 Flag to indicate that the cluster function implementation of \code{submitJob} can not directly handle \code{\link{JobCollection}} objects.
 If set to \code{FALSE}, the \code{\link{JobCollection}} is serialized to the file system before submitting the job.}
 
-\item{scheduler.delay}{[\code{numeric(1)}]\cr
+\item{scheduler.latency}{[\code{numeric(1)}]\cr
 Time to sleep after important interactions with the scheduler to ensure a sane state.
-Currently only user after \code{\link{submitJobs}}.}
+Currently only triggered after calling \code{\link{submitJobs}}.}
+
+\item{fs.latency}{[\code{numeric(1)}]\cr
+Expected maximum latency of the file system, in seconds.
+Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
+access files and directories.
+Usually safe to set to \code{NA} which disables the expensive heuristic if you are working on a local file system.}
 
 \item{hooks}{[\code{list}]\cr
 Named list of functions which will we called on certain events like \dQuote{pre.submit} or \dQuote{post.sync}.
