+++ man/makeClusterFunctionsOpenLava.Rd
@@ -5,7 +5,7 @@
 \title{ClusterFunctions for OpenLava}
 \usage{
 makeClusterFunctionsOpenLava(template = findTemplateFile("openlava"),
-  scheduler.delay = 1)
+  scheduler.latency = 1, fs.latency = 65)
 }
 \arguments{
 \item{template}{[\code{character(1)}]\cr
@@ -20,9 +20,15 @@ the first one found:
   \item \dQuote{openlava.default.tmpl} in the package installation directory in the subfolder \dQuote{templates} (this probably does not exist for your cluster system or needs adaptation).
 }}
 
-\item{scheduler.delay}{[\code{numeric(1)}]\cr
+\item{scheduler.latency}{[\code{numeric(1)}]\cr
 Time to sleep after important interactions with the scheduler to ensure a sane state.
-Currently only user after \code{\link{submitJobs}}.}
+Currently only triggered after calling \code{\link{submitJobs}}.}
+
+\item{fs.latency}{[\code{numeric(1)}]\cr
+Expected maximum latency of the file system, in seconds.
+Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
+access files and directories.
+Usually safe to set to \code{NA} which disables the expensive heuristic if you are working on a local file system.}
 }
 \value{
 [\code{\link{ClusterFunctions}}].
