### 2017-02-13 mllg/batchtools#88 @mllg
You are right, I am too restrictive regarding variable names. I always forget about the infix operators.

The reasoning for *not* storing the variable inside the RDS is that you have to read all files just to get the names, i.e. each time you call `batchExport`.

`make.names()` is not bijective, I would need something like a base32 encoder/decoder.

I guess I'll just store the real names in the registry and use a random string for the file names.
