+++ R/zzz.R
@@ -25,6 +25,7 @@
 #' @importFrom progress progress_bar
 #' @importFrom rappdirs user_config_dir
 #' @importFrom stats runif predict pexp
+#' @importFrom base64url base32_encode base32_decode
 "_PACKAGE"
 
 #' @title Deprecated function in the batchtools package
