+++ R/Export.R
@@ -34,14 +34,14 @@
 #' print(exports)
 batchExport = function(export = list(), unexport = character(0L), reg = getDefaultRegistry()) {
   assertRegistry(reg, writeable = TRUE)
-  assertList(export, names = "strict")
+  assertList(export, names = "named")
   assertCharacter(unexport, any.missing = FALSE, min.chars = 1L)
 
   path = file.path(reg$file.dir, "exports")
 
   if (length(export) > 0L) {
     nn = names(export)
-    fn = file.path(path, sprintf("%s.rds", nn))
+    fn = file.path(path, mangle(nn))
     found = file.exists(fn)
     if (any(!found))
       info("Exporting new objects: '%s' ...", stri_flatten(nn[!found], "','"))
@@ -51,7 +51,7 @@ batchExport = function(export = list(), unexport = character(0L), reg = getDefau
   }
 
   if (length(unexport) > 0L) {
-    fn = file.path(path, sprintf("%s.rds", unexport))
+    fn = file.path(path, mangle(unexport))
     found = file.exists(fn)
     if (any(found))
       info("Un-exporting exported objects: '%s' ...", stri_flatten(unexport[found], "','"))
@@ -59,5 +59,14 @@ batchExport = function(export = list(), unexport = character(0L), reg = getDefau
   }
 
   fns = list.files(path, pattern = "\\.rds")
-  invisible(data.table(name = stri_sub(fns, to = -5L), uri = file.path(path, fns)))
+  invisible(data.table(name = unmangle(fns), uri = file.path(path, fns)))
+}
+
+
+mangle = function(x) {
+  sprintf("%s.rds", base32_encode(x, use.padding = FALSE))
+}
+
+unmangle = function(x) {
+  base32_decode(stri_sub(x, to = -5L), use.padding = FALSE)
 }
