+++ appveyor.yml
@@ -20,7 +20,7 @@ environment:
 
 build_script:
   - travis-tool.sh install_deps
-  - travis-tool.sh install_github mllg/checkmate
+  - travis-tool.sh install_github mllg/checkmate mllg/base64ur
 
 test_script:
   - travis-tool.sh run_tests
