+++ NAMESPACE
@@ -123,6 +123,8 @@ import(data.table)
 import(stringi)
 import(utils)
 importFrom(R6,R6Class)
+importFrom(base64url,base32_decode)
+importFrom(base64url,base32_encode)
 importFrom(brew,brew)
 importFrom(digest,digest)
 importFrom(progress,progress_bar)
