### 2017-02-13 mllg/batchtools#88 @mllg
I've changed my mind. I think I will go with base32. See here: https://github.com/mllg/batchtools/pull/89
