### 2017-02-13 mllg/batchtools#88 @HenrikBengtsson
This is good news.

> The reasoning for not storing the variable inside the RDS is that you have to read all files just to get the names, i.e. each time you call `batchExport`.

I didn't consider that batchtools might need to see what's already exported.

FYI, I've used [`utils::URLencode()` / `utils::URLdecode()` in future.BatchJobs as a workaround](https://github.com/HenrikBengtsson/future.BatchJobs/blob/0.13.1/R/BatchJobsFuture-class.R#L436-L467) for a long time now and it's been working very well. Haven't had any issues. My point is, I think your approach will work without surprises.
