### 2017-02-14 mllg/batchtools#85 @HenrikBengtsson
No worries, I'm happy to help out getting this package as stable as possible.

So, good news. Today I've put batchtools 0.9.1-9000 (commit 46e2bfee) for some real-world serious testing on our TOEQUE / PBS cluster.  I did this via an early-version of future.batchtools which internally utilizes the batchtools package.  This allowed me to run through the same system tests that I also run sequential, plain futures (on top of parallel package), and future.BatchJobs.  Several of these tests also asserts identical results (regardless of computation backend). The tests runs for hours. **I get all OK in these tests**.
