### 2017-02-14 mllg/batchtools#85 @HenrikBengtsson
Feel free to close this issue whenever your done.
