### 2017-02-14 c0d94d0
hotfix for old registries
