### 2017-02-15 #135 @HenrikBengtsson
UPDATES:

Additional feedback based on some serious testing on real compute clusters:

* https://github.com/mllg/batchtools/issues/85
* https://github.com/mllg/batchtools/issues/87
* https://github.com/mllg/batchtools/issues/88

which @mmlg is addressing.

I've now successfully tested batchtools 0.9.1-9000 with:

* `makeClusterFunctionsInteractive(external = TRUE)` on multiple system including the below two clusters
* `makeClusterFunctionsTORQUE()` on a large TORQUE / PBS cluster running RedHat 6.6 and R 3.3.2
* `makeClusterFunctionsSGE()`: on a large Son of Grid Engine (SGE) cluster running Centos 7.3 and R 3.3.2

using some rather extensive real-world computations.

I'm extremely pleased with the experience I now have with batchtools.  **It's a very nice piece of work and extremely valuable!**  Thank you.

As soon as the developer's version batchtools 0.9.1-9000 is wrapped up and the updated version is submitted to CRAN, I'll be happy to recommend anyone to start using it for their HPC needs.  I also feel comfortable enough about this upcoming version of batchtools, that'll be recommending existing BatchJobs users to migrate over to batchtools (as the authors have suggested for a while now).

I hereby strongly recommend that the JOSS paper should be accepted.
