### 2017-02-15 09b0cbf
native routines
