### 2017-02-15 d82474f
fixed updateRegistry
