### 2017-02-15 420c52a
Docker fixes (#90)
