### 2017-02-15 #135 @HenrikBengtsson
Also, @arfon, all the checkboxes at the top can be ticked. I don't have the permission to do it.
