### 2017-02-15 9b1dee6
rework file handling
