+++ tests/testthat/test_sweepRegistry.R
@@ -10,14 +10,14 @@ test_that("sweepRegistry", {
   writeRDS(makeJobCollection(1, reg = reg), file.path(reg$file.dir, "jobs", "test.rds"))
 
   expect_data_table(reg$resources, nrow = 2)
-  expect_character(list.files(getLogPath(reg)), len = 2L)
+  expect_character(list.files(reg$path$log.dir), len = 2L)
   expect_character(list.files(file.path(reg$file.dir, "jobs"), pattern = "\\.rds$"), len = 1L + (array.jobs && reg$cluster.functions$store.job) * 2L)
   expect_character(list.files(file.path(reg$file.dir, "jobs"), pattern = "\\.job$"), len = (batchtools$debug && array.jobs) * 2L)
 
   expect_true(sweepRegistry(reg))
 
   expect_data_table(reg$resources, nrow = 1)
-  expect_character(list.files(getLogPath(reg)), len = 1L)
+  expect_character(list.files(reg$path$log.dir), len = 1L)
   if (reg$cluster.functions$store.job)
     expect_character(list.files(file.path(reg$file.dir, "jobs")), len = 0L)
   checkTables(reg)
