+++ R/batchMapResults.R
@@ -47,12 +47,14 @@ batchMapResults = function(fun, ids = NULL, ..., more.args = list(), target, sou
   if (nrow(target$status) > 0L)
     stop("Target registry 'target' must be empty")
 
-  more.args = c(list(..file.dir = source$file.dir, ..fun = fun), more.args)
+  fns = source$path$results(ids)
+  names(fns) = ids$job.id
+  more.args = c(list(..fn = fns, ..fun = fun), more.args)
   args = c(list(..id = ids$job.id), list(...))
 
   batchMap(batchMapResultsWrapper, args = args, more.args = more.args, reg = target)
 }
 
-batchMapResultsWrapper = function(..fun, ..file.dir, ..id, ...) {
-  ..fun(readRDS(getResultFiles(..file.dir, ..id)), ...)
+batchMapResultsWrapper = function(..fun, ..fn, ..id, ...) {
+  ..fun(readRDS(..fn[[..id]]), ...)
 }
