+++ R/removeExperiments.R
@@ -23,7 +23,7 @@ removeExperiments = function(ids = NULL, reg = getDefaultRegistry()) {
     info("Cleaning up %i job definitions ...", length(i))
     reg$defs = reg$defs[-i]
   }
-  fns = getResultFiles(reg$file.dir, ids$job.id)
+  fns = reg$path$results(ids)
   file.remove(fns[file.exists(fns)])
 
   sweepRegistry(reg)
