+++ tests/testthat/test_showLog.R
@@ -33,7 +33,7 @@ test_that("empty log files", {
   submitAndWait(reg)
 
   # overwrite log file
-  log.file = getLogFiles(reg$file.dir, reg$status[1, job.hash], reg$status[1, log.file])
+  log.file = reg$path$logs(reg$status[1])
   expect_true(file.remove(log.file))
   expect_true(file.create(log.file))
 

[…]
