### 2017-02-15 mllg/batchtools#87 @mllg
I now use base32 encoded file names for exports which should fix this. I will do the same thing for other user provided file names to make this consistent (i.e., problems and algorithms).
