### 2017-02-15 mllg/batchtools#88 @mllg
base32 encoding is merged. Should now work with all object names.
