### 2017-02-15 85e6646
dirty hack to avoid R6/locked bindings
