### 2017-02-15 0c6b449
use base32 for probs and algos
