+++ tests/testthat/test_addAlgorithm.R
@@ -6,7 +6,7 @@ test_that("addAlgorithm", {
   expect_is(algo, "Algorithm")
   expect_equal(algo$name, "a1")
   expect_function(algo$fun)
-  expect_file(getAlgorithmURI(reg, algo$name))
+  expect_file(reg$path$algorithms(algo$name))
 
   prob = addProblem(reg = reg, "p1", data = iris, fun = function(job, data) nrow(data))
   algo = addAlgorithm(reg = reg, "a2", fun = function(...) NULL)
@@ -17,8 +17,8 @@ test_that("addAlgorithm", {
   expect_integer(reg$status$job.id, len = 2L)
   expect_set_equal(levels(reg$defs$algorithm), "a2")
   expect_set_equal(getAlgorithmIds(reg), "a2")
-  expect_false(file.exists(getAlgorithmURI(reg, "a1")))
-  expect_true(file.exists(getAlgorithmURI(reg, "a2")))
+  expect_false(file.exists(reg$path$algorithms("a1")))
+  expect_true(file.exists(reg$path$algorithms("a2")))
   expect_set_equal(as.character(getJobPars(reg = reg)$algorithm), "a2")
   checkTables(reg)
 })
