### 2017-02-16 mllg/batchtools#88 @HenrikBengtsson
Awesome.  I've verified this with `makeClustersFunctionInteractive(external = TRUE)` and  `makeClustersFunctionTORQUE()`.
