### 2017-02-16 442bc78
use base32 for prob/algo; cleanup file handling
