+++ R/resetJobs.R
@@ -20,7 +20,7 @@ resetJobs = function(ids = NULL, reg = getDefaultRegistry()) {
   info("Resetting %i jobs in DB ...", nrow(ids))
   cols = c("submitted", "started", "done", "error", "memory", "resource.id", "batch.id", "log.file", "job.hash")
   reg$status[ids, (cols) := list(NA_real_, NA_real_, NA_real_, NA_character_, NA_real_, NA_integer_, NA_character_, NA_character_, NA_character_), on = "job.id"]
-  fns = getResultFiles(reg$file.dir, ids$job.id)
+  fns = getResultFiles(reg, ids)
   file.remove(fns[file.exists(fns)])
 
   sweepRegistry(reg)
