+++ R/helpers.R
@@ -164,3 +164,8 @@ chsetdiff = function(x, y) {
   # Note: assumes that x has no duplicates
   x[chmatch(x, y, 0L) == 0L]
 }
+
+chintersect = function(x, y) {
+  # Note: assumes that x has no duplicates
+  x[chmatch(y, x, 0L)]
+}
