+++ R/doJobCollection.R
@@ -43,7 +43,7 @@ doJobCollection.JobCollection = function(jc, output = NULL) {
     updates = data.table(job.id = jc$jobs$job.id, started = now, done = now,
       error = stri_trunc(stri_trim_both(sprintf(msg, ...)), 500L, " [truncated]"),
       memory = NA_real_, key = "job.id")
-    writeRDS(updates, file = getUpdateFiles(jc$file.dir, jc$job.hash, 0L))
+    writeRDS(updates, file = file.path(jc$file.dir, "updates", sprintf("%s.rds", jc$job.hash)))
     invisible(NULL)
   }
 
@@ -133,7 +133,7 @@ doJobCollection.JobCollection = function(jc, output = NULL) {
       update$error = stri_trunc(stri_trim_both(as.character(result)), 500L, " [truncated]")
     } else {
       catf("\n### [bt %s]: Job terminated successfully [batchtools job.id=%i]", now(), id)
-      writeRDS(result, file = getResultFiles(jc$file.dir, id))
+      writeRDS(result, file = getResultFiles(jc, id))
     }
     buf$add(i, update)
     buf$flush(jc)
@@ -165,7 +165,7 @@ UpdateBuffer = R6Class("UpdateBuffer",
       i = self$updates[!is.na(started) & (!written), which = TRUE]
       if (length(i) > 0L) {
         first.id = self$updates$job.id[i[1L]]
-        writeRDS(self$updates[i], file = getUpdateFiles(jc$file.dir, jc$job.hash, first.id))
+        writeRDS(self$updates[i], file = file.path(jc$file.dir, "updates", sprintf("%s-%i.rds", jc$job.hash, first.id)))
         set(self$updates, i, "written", TRUE)
       }
     },
