+++ tests/testthat/test_resetJobs.R
@@ -10,26 +10,26 @@ test_that("resetJobs", {
   )
   submitAndWait(reg, 1:3)
 
-  expect_true(file.exists(getLogFiles(reg$file.dir, reg$status[.(3), job.hash], reg$status[.(3), log.file])))
+  expect_true(file.exists(getLogFiles(reg, reg$status[.(3)])))
   expect_false(identical(reg$status$submitted, before$status$submitted))
-  expect_true(file.exists(getResultFiles(reg$file.dir, 1)))
+  expect_true(file.exists(getResultFiles(reg, 1)))
   expect_equal(dir.exists(file.path(reg$file.dir, "external", 1:3)), c(TRUE, FALSE, TRUE))
 
   resetJobs(1, reg = reg)
   expect_true(all.equal(before$status[1], reg$status[1]))
-  expect_false(file.exists(getResultFiles(reg$file.dir, 1)))
-  expect_true(file.exists(getResultFiles(reg$file.dir, 3)))
-  expect_true(file.exists(getLogFiles(reg$file.dir, reg$status[.(3L), job.hash], reg$status[.(3L), log.file])))
+  expect_false(file.exists(getResultFiles(reg, 1)))
+  expect_true(file.exists(getResultFiles(reg, 3)))
+  expect_true(file.exists(getLogFiles(reg, reg$status[.(3L)])))
   expect_equal(dir.exists(file.path(reg$file.dir, "external", 1:3)), c(FALSE, FALSE, TRUE))
-  expect_false(file.exists(getResultFiles(reg$file.dir, 1)))
-  expect_true(file.exists(getResultFiles(reg$file.dir, 3)))
+  expect_false(file.exists(getResultFiles(reg, 1)))
+  expect_true(file.exists(getResultFiles(reg, 3)))
 
   resetJobs(2:3, reg = reg)
   expect_data_table(reg$status, key = "job.id")
   expect_data_table(reg$defs, key = "def.id")
   expect_equivalent(before$status, reg$status)
-  expect_false(file.exists(getLogFiles(reg$file.dir, reg$status[.(3L), job.hash], NA_character_)))
-  expect_false(file.exists(getResultFiles(reg$file.dir, 3)))
+  expect_false(file.exists(getLogFiles(reg, reg$status[.(3L)])))
+  expect_false(file.exists(getResultFiles(reg, 3)))
   expect_equal(dir.exists(file.path(reg$file.dir, "external", 1:3)), c(FALSE, FALSE, FALSE))
 })
 
