### 2017-02-16 mllg/batchtools#81 @mllg
If this is not critical for the JOSS, I'd like to postpone this. Okay with you @HenrikBengtsson ?
