### 2017-02-16 mllg/batchtools#80 @mllg
Possible that the described mechanism is not sufficient. For some reasons systems are behaving differently than with the result files where a call to `list.files()` is required. Therefore I would like to stick with the current implementation until problems are reported. Ok with you @HenrikBengtsson ?
