### 2017-02-16 #135 @mllg
Thanks again @HenrikBengtsson for your thorough review. I will address the remaining small issues before Monday.

@arfon What is the supposed procedure to change the version number in this review (0.9.0) to match the version of the next package release (0.9.2)?
