### 2017-02-16 mllg/batchtools#81 @HenrikBengtsson
Certainly fine with me. This is a thing that shouldn't be rushed.
