### 2017-02-16 mllg/batchtools#80 @HenrikBengtsson
Just to be clear, I posted this issue based on the reading and not based on an experienced problem.   I must admit that this one slipped as I didn't really had time to dive into the details of your explanation.  It sounds like you're saying there's no/little risk for race conditions when it comes to JobCollection objects.

For other RDS files: Since you're using the RDS file format (and compressed too), I think (=hope) that is a good enough protection for one processing reading a half complete RDS file that another process yet hasn't finished writing; I hope it gives a read error.  The save to *.rds.tmp and then rename to *.rds was my idea to lower the risk for this even further since the reader would never see a half-baked file (only a full file or not a file at all).  If you're using `writeRDS()` everywhere, it sounds like an easy task to just add there.

But, I fully trust you here, especially since I only know so much about underlying design.  Feel free to ignore this issue until it really becomes an issue, if at all.
