### 2017-02-16 #135 @HenrikBengtsson
Just to clarify, none of the remaining batchtools issues should prevent the acceptance of the article; they can be dealt with later and are non-critical.
