### 2017-02-17 mllg/batchtools#81 @berndbischl
i have started to read up on this. and i am getting more and more worried.
can we please talk about this?
this concerns the validity of the results of everybody who computes with this.
