### 2017-02-17 mllg/batchtools#81 @berndbischl
https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/

https://rpubs.com/Jouni_Helske/225931
