### 2017-02-17 mllg/batchtools#81 @berndbischl
we are currently testing this out, but i am quite sure we will see what helske saw after he said:

```
Okay, what if we just forget the whole random seeding, and use consecutive seeds for example based on the thread id? This is what John suggests in his blog, and which I have seen suggested elsewhere as well. Problem with this approach is that if the different seed values are of form seedi=a+bseedi−1seedi=a+bseedi−1, the generated samples are not necessarily completely independent. I am no expert in random number generation, but my understanding is that most if not all the test regarding the quality of the RNG algorithms are concerned what happens in a single stream of random numbers. Instead of diving into theory, let us make a simple test in R.
```
