### 2017-02-17 mllg/batchtools#81 @berndbischl
i am running this now, which is the example in helske. and, naturally, i get the same results

```
load_all()

RNGkind("Marsaglia-Multicarry")
n = 10000
y1 = runif(n)
print(acf(y1))

makeRegistry(seed = 123L, file.dir = NA)

batchMap(function(x) {
  RNGkind("Marsaglia-Multicarry")
  runif(1)
}, 1:n)
submitJobs()
y2 = as.numeric(reduceResultsList())
print(acf(y2))
```

plots look really like in helske's link. i can see no worrysome auto-correlation for the mersenne, which is the default in R, but for  Marsaglia-Multicarry it looks very bad.

IMHO we really need to do something here, soon. This is completely intransparatrent to the user, and most people do not know enough about this, potentially including us. But nobody can debug this and discover potential problems if they appear on this front.
