### 2017-02-17 mllg/batchtools#81 @HenrikBengtsson
I'm using the L'Cuyer RNG stream approach, which, honestly, I use based on a fairly large portion of trust in how it's done in / chosen for the parallel package.

FYI, in [R Journal Volume 8/1, August 2016](https://journal.r-project.org/archive/2016-1/), there's article 'CryptRndTest: An R Package for Testing the Cryptographic Randomness' [[pdf]](https://journal.r-project.org/archive/2016-1/demirhan-bitirim.pdf) by Haydar Demirhan and Nihan Bitirim, which might be useful.  Their [CryptRndTest](https://cran.r-project.org/package=CryptRndTest) package is on CRAN.  I haven't more than skimmed through that paper.
