### 2017-02-17 mllg/batchtools#81 @berndbischl
> I'm using the L'Cuyer RNG stream approach, which, honestly, I use based on a fairly large portion of trust in how it's done in / chosen for the parallel package.

i looke at your code example above, again. thats what you are referring to.
am I making a mistake, or are you doing EXACTLY the same done in batchtools, with the only exception that you generate the seed in a special way?
