### 2017-02-17 mllg/batchtools#81 @berndbischl
(and i am pretty sure they have thought about the pretty hard for the parallel package, not distrusting that at all)
