### 2017-02-18 mllg/batchtools#81 @HenrikBengtsson
When iterating over elements `x`, you can do something like:
```r
RNGkind("L'Ecuyer-CMRG")
seed <- .GlobalEnv$.Random.seed
for (ii in seq_along(x)) {
   seed <- nextRNGStream(seed)
   job.seed <- nextRNGSubStream(seed)
   create_job_ii_with_seed(id = ii, x[[ii]], ..., seed = job.seed)
}
```
You don't want to just use `job.seed <- seed`, because then if job `ii` would call `seed_ii_next <- nextRNGStream(seed_ii)` in recursive calls, then `seed_ii_next == seed_jj` where `jj = ii + 1` (== the seed for job `ii+1`).  At least this is my interpretation on how to use `nextRNGStream()` and `nextRNGSubStream()`.

I'm basically using this approach in my (upcoming) `future_lapply()` which you  can find in https://github.com/HenrikBengtsson/future/blob/986eca/R/future_lapply.R, but it's a bit convoluted with other code too.
