### 2017-02-18 mllg/batchtools#81 @HenrikBengtsson
Forgot to say, I don't see any usage of `nextRNGStream()` in batchtools, which uses https://github.com/mllg/batchtools/blob/master/R/helpers.R#L145-L150
