### 2017-02-18 #135 @arfon
> Just to clarify, none of the remaining batchtools issues should prevent the acceptance of the article; they can be dealt with later and are non-critical.

Thanks for the clarification.

> @arfon What is the supposed procedure to change the version number in this review (0.9.0) to match the version of the next package release (0.9.2)?

Just to let me know - I've made the change at the top of the thread here.

@mllg - Once you're ready, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
