### 2017-02-19 4b4e5de
remove stray autocomplete leftovers (fixes #93)
