+++ R/removeRegistry.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' All files will be erased from the file system, including all results.
-#' If you wish to remove only intermediate files, use \code{\link{sweepRegistry}}.<`3`>
+#' If you wish to remove only intermediate files, use \code{\link{sweepRegistry}}.
 #'
 #' @param wait [\code{numeric(1)}]\cr
 #'   Seconds to wait before proceeding. This is a safety measure to not
