### 2017-02-19 c44cc6c
added new funs to pkgdown [ci skip]
