### 2017-02-20 9530cf1
fix loading of old registries
