### 2017-02-20 21a11b4
re-enabled parallelMap tests
