### 2017-02-20 4488add
disabled some tests for solaris / broken system2
