### 2017-02-20 #135 @mllg
I've published a new release on [CRAN](https://cran.r-project.org/package=batchtools) and [Github/Zenodo](https://zenodo.org/record/293835).

DOI is <http://doi.org/10.5281/zenodo.293835>.
