### 2017-02-22 5b682b5
increased version number after release
