### 2017-02-22 #135 @arfon
Many thanks for your reviews @masalmon and @HenrikBengtsson ✨

@mllg - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00135 :zap: 🚀 💥
